Split optimize submit into focused modules
Keep the existing CLI and import facade while moving the implementation into smaller CI support modules for maintainability.
