+++ ci/optimize_submit.py
@@ -41,3941 +41,60 @@ Env vars (all optional, CLI flags take precedence):
   HF_TOKEN                           HuggingFace token (gated models)
   SAFE_OPTIMIZE_WORKSPACE            override default 'core42-hyperloom'
   SAFE_OPTIMIZE_VOLUME               override default '/wekafs'
+
+Implementation note:
+  This file is intentionally kept as the stable CLI/import facade. The
+  implementation lives under ci/optimize_submit_lib/ so individual modules stay
+  reviewable while workflows keep running python3 optimize_submit.py.
 """
 
 from __future__ import annotations
 
 import argparse
-import json
 import logging
 import os
 import random
-import re
-import shutil
 import sys
 import time
-from dataclasses import asdict, dataclass, field
-from datetime import datetime, timezone, timedelta
 from pathlib import Path
 
-import requests
-
+# Keep local sibling imports (model_compat, optimize_submit_lib) available when
+# the script is executed directly from ci/ or imported by tests.
 sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))
-import model_compat  # noqa: E402  (local sibling module: shared compat rules)
+
+from optimize_submit_lib import artifacts as _artifacts
+from optimize_submit_lib import config as _config
+from optimize_submit_lib import detect as _detect
+from optimize_submit_lib import flow as _flow
+from optimize_submit_lib import hf_client as _hf_client
+from optimize_submit_lib import manifest as _manifest
+from optimize_submit_lib import records as _records
+from optimize_submit_lib import safe_client as _safe_client
+
+for _module in (
+    _config,
+    _hf_client,
+    _detect,
+    _safe_client,
+    _records,
+    _artifacts,
+    _flow,
+    _manifest,
+):
+    globals().update({k: v for k, v in vars(_module).items() if not k.startswith("__")})
 
 log = logging.getLogger("optimize-submit")
 
 
-# ── Defaults ────────────────────────────────────────────────────────────────────
-
-DEFAULT_API_URL = "https://core42.primus-safe.amd.com"
-# Two-workspace split due to conflicting K8s constraints: register =
-# core42-hyperloom (has /wekafs RWX for downloads; shared volume is readable by
-# all sandboxes); submit = core42-sandbox (only Sandbox-scoped workspaces can
-# host the optimization task; core42-hyperloom rejects with Primus.00003).
-# Needs SaFE selectLocalPath path-accessible fallback when submit != register,
-# else submit_task 400s; workaround: set --submit-workspace == --register-workspace.
-DEFAULT_REGISTER_WORKSPACE = "core42-hyperloom"
-DEFAULT_SUBMIT_WORKSPACE = "core42-sandbox"
-DEFAULT_VOLUME = "/wekafs"
-# core42 is MI300X; override the Claw prompt-builder's wrong MI355X default so
-# the prompt and TP policy use the right arch. Tool source paths are
-# deliberately NOT pinned: install.sh clones writable per-session copies, and
-# pinning the shared read-only /wekafs/hyperloom/InferenceX caused
-# "Read-only file system" errors on first session. OOB/TraceLens are opt-in too.
-DEFAULT_GPU_TYPE = "MI300X"
-DEFAULT_GPU_PROFILE = "mi300x"
-DEFAULT_KERNEL_BACKENDS = ["GEAK", "Claude Code", "Codex"]
-DEFAULT_MAX_HOURS = 6.0
-DEFAULT_TARGET_GAIN = 500.0
-DEFAULT_RESULTS_PATH = "$RESULT_DIR"
-DEFAULT_CONTEXT_RESERVE_TOKENS = 16
-# Models whose config.json ``max_position_embeddings`` is at or below this are
-# skipped outright: their context window is too small to be worth a sandbox slot
-# (independent of the per-workload ``context_too_short`` check).
-MIN_MAX_POSITION_EMBEDDINGS = 2048
-
-# Hardware facts from AMD Instinct datasheets. tp_thresholds_b is CI policy:
-# MI300X baseline (80/128/256B): <=80B->TP1, <=128B->TP2, <=256B->TP4,
-# >256B->TP8. mi325x/mi355x stay scaled by per-GPU HBM capacity.
-GPU_PROFILES = {
-    "mi300x": {
-        "gpu_type": "MI300X",
-        "llvm_target": "gfx942",
-        "hbm_gb": 192,
-        "hbm_bandwidth_tb_s": 5.3,
-        "tp_thresholds_b": (80, 128, 256),
-    },
-    "mi325x": {
-        "gpu_type": "MI325X",
-        "llvm_target": "gfx942",
-        "hbm_gb": 256,
-        "hbm_bandwidth_tb_s": 6.0,
-        "tp_thresholds_b": (43, 171, 341),
-    },
-    "mi355x": {
-        "gpu_type": "MI355X",
-        "llvm_target": "gfx950",
-        "hbm_gb": 288,
-        "hbm_bandwidth_tb_s": 8.0,
-        "tp_thresholds_b": (48, 192, 384),
-    },
-}
-
-_KERNEL_BACKEND_ALIASES = {
-    "geak": "GEAK",
-    "claude": "Claude Code",
-    "claude-code": "Claude Code",
-    "claude code": "Claude Code",
-    "codex": "Codex",
-    "cursor": "Cursor",
-}
-
-
-def normalize_gpu_profile(gpu_type: str | None, *, warn: bool = True) -> str | None:
-    """Return the GPU profile key when ``gpu_type`` maps to a known CI profile.
-
-    Args:
-        gpu_type (str | None): Free-form GPU type/alias.
-        warn (bool): Log a warning when a non-empty ``gpu_type`` is unknown.
-
-    Returns:
-        str | None: The matching :data:`GPU_PROFILES` key, or ``None`` when no
-        profile matches.
-    """
-    raw = (gpu_type or "").strip()
-    compact = re.sub(r"[^a-z0-9]", "", raw.lower())
-    for key, profile in GPU_PROFILES.items():
-        profile_type = str(profile["gpu_type"])
-        aliases = {
-            re.sub(r"[^a-z0-9]", "", key.lower()),
-            re.sub(r"[^a-z0-9]", "", profile_type.lower()),
-        }
-        if compact in aliases or compact.endswith(key):
-            return key
-    if warn and raw:
-        log.warning(
-            "unknown gpu_type=%r; using %s for TP policy only", raw, GPU_PROFILES[DEFAULT_GPU_PROFILE]["gpu_type"]
-        )
-    return None
-
-
-def canonical_gpu_type(gpu_type: str | None) -> str:
-    """Resolve a GPU type string to its canonical form.
-
-    Args:
-        gpu_type: Free-form GPU type/alias, or ``None``.
-
-    Returns:
-        The canonical GPU type from the matching profile, or the trimmed input
-        (falling back to ``DEFAULT_GPU_TYPE``) when no profile matches.
-    """
-    profile_key = normalize_gpu_profile(gpu_type, warn=False)
-    if profile_key:
-        return str(GPU_PROFILES[profile_key]["gpu_type"])
-    return (gpu_type or DEFAULT_GPU_TYPE).strip() or DEFAULT_GPU_TYPE
-
-
-# Canonical prompt prefix (ci/prompt_prefix.txt) — single source of truth,
-# also read by .github/workflows/optimize-submit.yml and used as the argparse
-# default so direct CLI invocations get the same prefix.
-_PROMPT_PREFIX_FILE = Path(__file__).resolve().parent / "prompt_prefix.txt"
-
-
-def _load_default_prompt_prefix() -> str:
-    """Resolve the default ``--prompt-prefix``: $SAFE_OPTIMIZE_PROMPT_PREFIX,
-    else ci/prompt_prefix.txt, else empty string.
-
-    Returns:
-        str: The resolved prompt prefix, or ``""`` when no source supplies one.
-    """
-
-    env_value = os.environ.get("SAFE_OPTIMIZE_PROMPT_PREFIX", "")
-    if env_value:
-        return env_value
-    try:
-        if _PROMPT_PREFIX_FILE.is_file():
-            return _PROMPT_PREFIX_FILE.read_text(encoding="utf-8")
-    except OSError:
-        # Prefix file unreadable; fall back to the empty prefix below.
-        pass
-    return ""
-
-
-# RDMA bnxt_re tar package staged on WekaFS — same path the validated
-# ci-config.yaml multi-node entries hand the agent. Kept as a module constant
-# so the orchestrator (Claw-direct) and this SaFE path stay in sync.
-_MULTINODE_BNXT_TAR = "/wekafs/primus/data/libbnxt/libbnxt_re-234.0.154.0.tar.gz"
-
-
-def _multinode_prompt_suffix(nodes: int, rayjob_image: str) -> str:
-    """RayJob topology block injected into the prompt when nodes > 1.
-
-    The SaFE tasks body has no node-count field, so multi-node is expressed the
-    same way the Claw-direct CI does: tell the agent to fan out to an N-node
-    RayJob with the exact topology. Returns "" for single-node (unchanged path).
-
-    Args:
-        nodes (int): Node count for the run; ``<= 1`` produces no suffix.
-        rayjob_image (str): Container image to advertise for the RayJob.
-
-    Returns:
-        str: The multi-node topology prompt block, or ``""`` for single-node.
-    """
-    if nodes <= 1:
-        return ""
-    return (
-        f"\n\nMulti-node run ({nodes} nodes): this model does NOT fit on a single "
-        f"8-GPU sandbox node. You CANNOT benchmark it directly in the sandbox -- "
-        f"launch it as a SaFE RayJob (Claw fan-out) and export NODES={nodes} "
-        f"before starting `inference_optimizer optimize` so the Coordinator sizes "
-        f"the cluster for {nodes}x8 GPUs (TP spans all nodes).\n"
-        f"Task submission ({nodes}-node):\n"
-        f"RayJob image: {rayjob_image}\n"
-        f"RayJob resource per node: CPU=96, GPU=8, memory=1024Gi, ephemeralStorage=400Gi\n"
-        f"RayJob node count: {nodes}\n"
-        f"env:\n"
-        f"- NODES={nodes}\n"
-        f"- PATH_TO_BNXT_TAR_PACKAGE={_MULTINODE_BNXT_TAR}\n"
-    )
-
-
-def parse_kernel_backends(raw: str | None) -> list[str]:
-    """Normalize user-facing kernel backend names for SaFE's API payload.
-
-    Splits on commas/semicolons, lowercases, and maps each alias to its
-    canonical SaFE name, de-duplicating while preserving order.
-
-    Args:
-        raw (str | None): Comma/semicolon-separated backend names, or None.
-
-    Returns:
-        list[str]: Canonical backend names, or the default list when ``raw`` is
-            empty or yields nothing.
-
-    Raises:
-        ValueError: If a token is not a recognised backend alias.
-    """
-
-    if not raw:
-        return list(DEFAULT_KERNEL_BACKENDS)
-    out: list[str] = []
-    for part in raw.replace(";", ",").split(","):
-        item = part.strip()
-        if not item:
-            continue
-        key = item.lower()
-        normalized = _KERNEL_BACKEND_ALIASES.get(key)
-        if normalized is None:
-            raise ValueError(f"unknown kernel backend {item!r}; expected one of geak, claude, codex, cursor")
-        if normalized not in out:
-            out.append(normalized)
-    return out or list(DEFAULT_KERNEL_BACKENDS)
-
-
-# Architectures well-supported by SGLang on ROCm 7.x.
-SGLANG_ARCHS: set[str] = {
-    "LlamaForCausalLM",
-    "LlamaForCausalLMWithVisualEncoder",
-    "Qwen2ForCausalLM",
-    "Qwen3ForCausalLM",
-    "Qwen2MoeForCausalLM",
-    "Qwen3MoeForCausalLM",
-    "MistralForCausalLM",
-    "MixtralForCausalLM",
-    "DeepseekV2ForCausalLM",
-    "DeepseekV3ForCausalLM",
-    "DeepseekV32ForCausalLM",
-    "GemmaForCausalLM",
-    "Gemma2ForCausalLM",
-    "Gemma3ForCausalLM",
-    "InternLM2ForCausalLM",
-    "InternLM3ForCausalLM",
-    "Phi3ForCausalLM",
-    "PhiForCausalLM",
-    "GPTBigCodeForCausalLM",
-    "FalconForCausalLM",
-    "ChatGLMModel",
-    # New architectures natively supported by sglang v0.5.12 (transformers 5.x)
-    # in the current sandbox image. Without these, detect_framework falls back
-    # to vLLM and the vllm-openai-rocm:v0.21.0 image (transformers
-    # <5) crashes at baseline ("does not recognize this architecture" /
-    # "TokenizersBackend does not exist"). Verified against the failing models'
-    # config.architectures.
-    "Gemma4ForConditionalGeneration",            # gemma-4 (dense + A4B MoE)
-    "Qwen3_5ForConditionalGeneration",           # Qwen3.5 / Qwen3.6 dense
-    "Qwen3_5MoeForConditionalGeneration",        # Qwen3.5 / Qwen3.6 A3B MoE
-    "Mistral3ForConditionalGeneration",          # Mistral3 / Ministral3
-    "NemotronHForCausalLM",                      # Nemotron-H (nemotron_h)
-    "Glm4ForCausalLM",                           # GLM-4 (glm4) dense
-    "Glm4MoeForCausalLM",                        # GLM-4.5 / 4.6 (glm4_moe)
-}
-
-# Architectures that require vLLM (Lightning Attention, sparse, or special quant).
-VLLM_REQUIRED_ARCHS: set[str] = {
-    "MiniMaxText01ForCausalLM",
-    "KimiForConditionalGeneration",
-    "KimiK25ForConditionalGeneration",
-}
-
-# Quantization types that require vLLM.
-VLLM_QUANT_TYPES: set[str] = {"mxfp4", "nvfp4", "int4", "gptq", "awq"}
-
-# Arch suffixes marking a generative LM; used to filter out embedding/encoder/
-# classifier models that sglang/vllm can't serve.
-GENERATIVE_ARCH_SUFFIXES: tuple[str, ...] = (
-    "ForCausalLM",
-    "ForConditionalGeneration",
-    "LMHeadModel",
-    "ForSeq2SeqLM",
-)
-
-
-def is_generative_arch(arch: str) -> bool:
-    """True if the HF arch is causal-LM-style inference; False for empty/unknown
-    (better to skip than waste a sandbox slot).
-
-    Args:
-        arch (str): HF ``architectures[0]`` class name.
-
-    Returns:
-        bool: True when ``arch`` ends with a known generative suffix.
-    """
-    if not arch:
-        return False
-    return any(arch.endswith(s) for s in GENERATIVE_ARCH_SUFFIXES)
-
-
-def _default_sglang_image() -> str:
-    """Return the default SGLang server image.
-
-    Returns:
-        The pinned ``profilerfix`` SGLang image whose patched
-        libamdhip64/libroctracer let rocprofiler capture kernels under
-        ``HipGraphLaunch`` (issue #352).
-    """
-    # profilerfix: patched libamdhip64/libroctracer so rocprofiler captures
-    # kernels under HipGraphLaunch (issue #352). Pre-profilerfix image (revert):
-    # lmsysorg/sglang:v0.5.12-rocm720-mi30x
-    return "harbor.core42.primus-safe.amd.com/sync/primussafe/sglang:v0.5.12-rocm720-mi30x-profilerfix"
-
-
-def _default_vllm_image() -> str:
-    """Return the default vLLM server image.
-
-    Returns:
-        The pinned ``profilerfix`` vLLM image (v0.21.0 / rocm720) whose patched
-        libamdhip64/libroctracer let rocprofiler capture kernels under
-        HipGraphLaunch (same profilerfix rationale as the SGLang image).
-    """
-    return "harbor.core42.primus-safe.amd.com/sync/primussafe/vllm-openai-rocm:v0.21.0-rocm720-profilerfix-20260627"
-
-
-# ── HuggingFace client ──────────────────────────────────────────────────────────
-
-
-class HuggingFaceClient:
-    """Minimal HF API client for model metadata + top-models discovery."""
-
-    BASE = "https://huggingface.co"
-
-    def __init__(self, token: str = "", timeout: int = 15, tokens: list[str] | None = None, seed: str = ""):
-        """Initialise the HF client session.
-
-        Supports a *pool* of HuggingFace tokens. Requests start on a token chosen
-        by hashing ``seed`` (so parallel CI jobs spread their first hit across the
-        pool) and, on HTTP 429 (Too Many Requests), transparently rotate to the
-        next token with exponential backoff before retrying. This mitigates the
-        bursty rate-limiting that hits the resolve/config.json endpoint when many
-        optimize jobs fan out at once.
-
-        Args:
-            token (str): Primary HuggingFace token for gated-model access.
-            timeout (int): Per-request timeout in seconds.
-            tokens (list[str] | None): Additional tokens to alternate with.
-            seed (str): Per-job string used to pick the starting token.
-        """
-        self.timeout = timeout
-        pool: list[str] = []
-        for t in [token, *(tokens or [])]:
-            if t and t not in pool:
-                pool.append(t)
-        self._tokens = pool
-        if pool:
-            self._idx = (hash(seed) % len(pool)) if seed else 0
-        else:
-            self._idx = 0
-        self._sess = requests.Session()
-        self._sess.headers["User-Agent"] = "hyperloom-optimize-submit/1.0"
-        self._apply_token()
-
-    def _apply_token(self) -> None:
-        """Set the Authorization header to the currently-selected token."""
-        if self._tokens:
-            tok = self._tokens[self._idx % len(self._tokens)]
-            self._sess.headers["Authorization"] = f"Bearer {tok}"
-        else:
-            self._sess.headers.pop("Authorization", None)
-
-    def _get(self, path: str) -> dict | list:
-        """GET a HuggingFace API path and return the parsed JSON body.
-
-        On HTTP 429 the client rotates to the next token in the pool and retries
-        with exponential backoff (capped at 30s); other errors raise immediately.
-
-        Args:
-            path (str): Path appended to the HF base URL.
-
-        Returns:
-            dict | list: The decoded JSON response.
-
-        Raises:
-            requests.HTTPError: If the response status indicates an error.
-        """
-        attempts = max(4, len(self._tokens) * 2)
-        last_exc: Exception | None = None
-        for i in range(attempts):
-            resp = self._sess.get(f"{self.BASE}{path}", timeout=self.timeout)
-            if resp.status_code == 429:
-                last_exc = requests.HTTPError(
-                    f"429 Client Error: Too Many Requests for url: {self.BASE}{path}", response=resp
-                )
-                if len(self._tokens) > 1:
-                    self._idx += 1
-                    self._apply_token()
-                time.sleep(min(2 ** i, 30))
-                continue
-            resp.raise_for_status()
-            return resp.json()
-        if last_exc is not None:
-            raise last_exc
-        raise requests.HTTPError(f"exhausted retries for url: {self.BASE}{path}")
-
-    def model_info(self, repo_id: str) -> dict:
-        """Fetch a repo's model metadata from the HF API.
-
-        Args:
-            repo_id (str): HuggingFace repo id (e.g. ``Qwen/Qwen3-8B``).
-
-        Returns:
-            dict: The model-info JSON (pipeline_tag, safetensors, tags, etc.).
-        """
-        return self._get(f"/api/models/{repo_id}")  # type: ignore[return-value]
-
-    def model_config(self, repo_id: str) -> dict:
-        """Fetch a repo's ``config.json`` from the HF resolve endpoint.
-
-        Args:
-            repo_id (str): HuggingFace repo id.
-
-        Returns:
-            dict: The parsed ``config.json`` contents.
-        """
-        return self._get(f"/{repo_id}/resolve/main/config.json")  # type: ignore[return-value]
-
-    def top_models(self, limit: int, min_params_b: float = 0.0) -> list[str]:
-        """Return top-N text-generation repos by downloads, optionally size-filtered.
-
-        Pool-then-filter: the listing API matches on tags only, so re-validate
-        per-repo on a generative architectures[0] suffix; failing that (or a
-        gated 401) → skip. The pipeline_tag != text-generation gate has been
-        removed so multimodal/other heads are allowed through.
-
-        Args:
-            limit (int): Maximum number of repos to return.
-            min_params_b (float): Minimum parameter count in billions; ``0``
-                disables the size filter.
-
-        Returns:
-            list[str]: Up to ``limit`` generative repo ids ordered by downloads.
-        """
-        pool_size = max(limit * 10, 100)
-        listing = self._get(f"/api/models?sort=downloads&direction=-1&limit={pool_size}&filter=text-generation")
-        repos: list[str] = []
-        for m in listing:  # type: ignore[union-attr]
-            if len(repos) >= limit:
-                break
-            repo = m.get("modelId") or m.get("id", "")
-            if not repo or "/" not in repo:
-                continue
-
-            try:
-                info = self.model_info(repo)
-            except Exception:
-                continue  # gated / network error
-
-            if min_params_b > 0:
-                total = (info.get("safetensors") or {}).get("total", 0)
-                if (total / 1e9) < min_params_b:
-                    continue
-
-            # Final gate: config.json reachable AND architectures[0] generative.
-            # Skip gated (401/403) or non-generative repos so the pool refills.
-            try:
-                cfg = self.model_config(repo)
-            except Exception as e:
-                log.info(
-                    "skip %s: config.json unreachable (%s) — likely a gated repo your HF_TOKEN hasn't been granted",
-                    repo,
-                    e,
-                )
-                continue
-            arch = (cfg.get("architectures") or [""])[0]
-            if not is_generative_arch(arch):
-                log.info("skip %s: arch=%s is non-generative", repo, arch)
-                continue
-
-            repos.append(repo)
-        return repos
-
-
-# ── Auto-detection ──────────────────────────────────────────────────────────────
-
-
-@dataclass
-class DetectedConfig:
-    """Auto-detected launch configuration for a model.
-
-    Attributes:
-        arch (str): HF ``architectures[0]`` class name.
-        framework (str): Chosen serving framework (``sglang`` / ``vllm``).
-        precision (str): Detected precision tag (e.g. ``FP8`` / ``INT4``).
-        tp (int): Tensor-parallel size.
-        concurrency (int): Benchmark concurrency.
-        image (str): Container image to run.
-        params_b (float): Parameter count in billions.
-    """
-
-    arch: str
-    framework: str
-    precision: str
-    tp: int
-    concurrency: int
-    image: str
-    params_b: float
-    max_context_tokens: int
-    # Raw ``config.json`` ``max_position_embeddings`` (0 when absent); used for
-    # the absolute small-context skip floor (``MIN_MAX_POSITION_EMBEDDINGS``).
-    max_position_embeddings: int = 0
-    # Raw ``config.json`` dict, used by the shared model_compat pre-flight.
-    raw_config: dict = field(default_factory=dict)
-
-
-def _quant_type(config: dict) -> str:
-    """Read the quantization tag from HF config.json (vendors disagree on the
-    field name). Priority: quant_algo (NVIDIA modelopt — quant_method there is
-    just the tool name) > quant_type > quantization_type > quant_method
-    (current de-facto standard) > method. First non-empty wins, lowercased.
-
-    Args:
-        config (dict): A HF ``config.json`` dict.
-
-    Returns:
-        str: The lowercased quantization tag, or ``""`` when none is present.
-    """
-    quant = config.get("quantization_config") or {}
-    raw = (
-        quant.get("quant_algo")
-        or quant.get("quant_type")
-        or quant.get("quantization_type")
-        or quant.get("quant_method")
-        or quant.get("method")
-        or ""
-    )
-    return raw.lower() if isinstance(raw, str) else ""
-
-
-def detect_framework(config: dict) -> str:
-    """Choose the serving framework for a model from its config.
-
-    vLLM is selected for architectures that require it or for quantization
-    types it handles better; SGLang is used for known-good architectures;
-    otherwise vLLM is the broader-support fallback.
-
-    Args:
-        config (dict): A HF ``config.json`` dict.
-
-    Returns:
-        str: ``"vllm"`` or ``"sglang"``.
-    """
-    arch = (config.get("architectures") or [""])[0]
-    qt = _quant_type(config)
-    if arch in VLLM_REQUIRED_ARCHS:
-        return "vllm"
-    if any(q in qt for q in VLLM_QUANT_TYPES):
-        return "vllm"
-    if arch in SGLANG_ARCHS:
-        return "sglang"
-    log.warning("unknown architecture %r — defaulting to vllm (broader support)", arch)
-    return "vllm"
-
-
-def detect_precision(config: dict) -> str:
-    """Detect the serving precision from a model's quantization tag.
-
-    Args:
-        config (dict): A HF ``config.json`` dict.
-
-    Returns:
-        str: One of ``FP8`` / ``FP4`` / ``INT4``, defaulting to ``FP8`` for
-            unquantized models on MI300X.
-    """
-    qt = _quant_type(config)
-    if "fp8" in qt:
-        return "FP8"
-    if "mxfp4" in qt:
-        return "FP4"
-    if "nvfp4" in qt:
-        return "FP4"
-    if "int4" in qt:
-        return "INT4"
-    if "gptq" in qt:
-        return "INT4"
-    if "awq" in qt:
-        return "INT4"
-    return "FP8"  # unquantized default for MI300X
-
-
-def detect_param_count(hf_info: dict, config: dict) -> float:
-    """Estimate a model's parameter count in billions.
-
-    Prefers the exact ``safetensors.total`` count from HF metadata; otherwise
-    approximates from hidden size, layer count, and vocab size.
-
-    Args:
-        hf_info (dict): The HF model-info JSON.
-        config (dict): The HF ``config.json`` dict.
-
-    Returns:
-        float: Parameter count in billions, or 0.0 when it cannot be estimated.
-    """
-    total = (hf_info.get("safetensors") or {}).get("total", 0)
-    if total:
-        return total / 1e9
-    h = config.get("hidden_size", 0)
-    n = config.get("num_hidden_layers", 0)
-    vocab = config.get("vocab_size", 0)
-    if h and n:
-        return (12 * h * h * n + vocab * h) / 1e9
-    return 0.0
-
-
-def detect_max_context_tokens(config: dict) -> int:
-    """Return the model context length from HF config.json when present.
-
-    Args:
-        config (dict): A HF ``config.json`` dict.
-
-    Returns:
-        int: The smallest positive context-length field found, or ``0`` when
-        none is present.
-    """
-    candidates = []
-    for key in ("max_position_embeddings", "max_sequence_length", "n_positions", "seq_length"):
-        value = config.get(key)
-        if isinstance(value, (int, float)) and value > 0:
-            candidates.append(int(value))
-    return min(candidates) if candidates else 0
-
-
-def context_too_short(
-    max_context_tokens: int,
-    isl: int,
-    osl: int,
-    reserve_tokens: int = DEFAULT_CONTEXT_RESERVE_TOKENS,
-) -> bool:
-    """Return whether the model context cannot fit the requested workload.
-
-    Args:
-        max_context_tokens: Model's maximum context length; ``<= 0`` means
-            unknown, in which case the check is skipped.
-        isl: Input sequence length.
-        osl: Output sequence length.
-        reserve_tokens: Headroom kept free beyond input + output.
-
-    Returns:
-        ``True`` when the context is known and smaller than
-        ``isl + osl + reserve_tokens``; otherwise ``False``.
-    """
-    if max_context_tokens <= 0:
-        return False
-    return max_context_tokens < (isl + osl + reserve_tokens)
-
-
-def detect_tp(params_b: float, precision: str = "BF16", gpu_type: str | None = None) -> int:
-    """Pick tensor parallelism from param count and GPU profile. precision is
-    kept for API compatibility but unused — CI policy stays predictable across
-    precision variants of the same model family.
-
-    Args:
-        params_b (float): Parameter count in billions.
-        precision (str): Kept for API compatibility; unused.
-        gpu_type (str | None): GPU type used to select the profile thresholds.
-
-    Returns:
-        int: The tensor-parallel size (1, 2, 4, or 8).
-    """
-    if params_b <= 0:
-        return 1
-    profile_key = normalize_gpu_profile(gpu_type) or DEFAULT_GPU_PROFILE
-    profile = GPU_PROFILES[profile_key]
-    tp1_max, tp2_max, tp4_max = profile["tp_thresholds_b"]
-    if params_b <= tp1_max:
-        return 1
-    if params_b <= tp2_max:
-        return 2
-    if params_b <= tp4_max:
-        return 4
-    return 8
-
-
-def detect_concurrency(tp: int, framework: str) -> int:
-    """Pick a benchmark concurrency from tensor-parallel size and framework.
-
-    CI policy is now a single fixed concurrency of 64 across every framework
-    and TP size, so benchmark load stays directly comparable between models
-    regardless of how they are sharded or served. ``tp`` / ``framework`` are
-    kept for API compatibility.
-
-    Args:
-        tp (int): Tensor-parallel size.
-        framework (str): Serving framework (``vllm`` / ``sglang``).
-
-    Returns:
-        int: The chosen concurrency level (always ``64``).
-    """
-    return 64
-
-
-def _sglang_image_for(repo_id: str = "") -> str:
-    """Pick the sglang image, honoring per-model baseline-arch needs.
-
-    Default is the v0.5.12 profilerfix image. Older branches carried a MiMo-V2
-    exception that pinned ``primussafe/sglang:v0.5.11-rocm720-mi30x-mimo-
-    profilerfix`` because the then-current profilerfix image did not register
-    ``MiMoV2ForCausalLM``. The v0.5.12 profilerfix image now includes the MiMo
-    model files, so MiMo should use the same current base as other SGLang
-    models. It is still paired with ``--attention-backend triton`` (injected in
-    ``_workload_envs.materialize_config_with_envs``) to dodge the aiter
-    attention CUDA-graph-capture SIGABRT.
-
-    Args:
-        repo_id (str): Model repo id, matched on its basename for overrides.
-
-    Returns:
-        str: The default sglang image. ``repo_id`` is accepted for backward
-        compatibility (and future per-model overrides) but no longer triggers a
-        MiMo-specific image; MiMo runs on the same default base.
-    """
-    return _default_sglang_image()
-
-
-def _vllm_image_for(repo_id: str = "") -> str:
-    """Pick the vLLM image, honoring per-model baseline-arch needs.
-
-    Default is the standard vLLM image. Gemma-4 (e.g. google/gemma-4-26B-A4B-it)
-    is the exception: the stock vLLM build does not serve the gemma-4 arch, so it
-    needs the dedicated gemma4 image. Matched on the repo basename so it fires
-    for the HF repo id regardless of org casing. Only consulted on the vLLM path
-    (sglang is unaffected).
-
-    Args:
-        repo_id (str): Model repo id, matched on its basename for overrides.
-
-    Returns:
-        str: The gemma4 vLLM image for gemma-4 repos, else the default vLLM image.
-    """
-    basename = (repo_id or "").split("/")[-1].lower()
-    if "gemma-4" in basename or "gemma4" in basename:
-        return "harbor.core42.primus-safe.amd.com/sync/vllm-openai-rocm:gemma4"
-    return _default_vllm_image()
-
-
-def detect_image(framework: str, repo_id: str = "") -> str:
-    """Select the server image for a framework and model.
-
-    Args:
-        framework: Serving framework (``vllm`` / ``sglang``).
-        repo_id: Model repo id, used to honor per-model image overrides.
-
-    Returns:
-        The vLLM image chosen by :func:`_vllm_image_for` for ``vllm`` (gemma-4
-        gets a dedicated image); otherwise the SGLang image chosen by
-        :func:`_sglang_image_for`.
-    """
-    return _vllm_image_for(repo_id) if framework == "vllm" else _sglang_image_for(repo_id)
-
-
-def auto_detect(hf: HuggingFaceClient, repo_id: str, gpu_type: str | None = None) -> DetectedConfig | None:
-    """Derive a benchmark configuration from a model's HF metadata.
-
-    Fetches model info and ``config.json`` and infers framework, precision,
-    tensor parallelism, concurrency, image, and context limits.
-
-    Args:
-        hf: Hugging Face client used to fetch metadata.
-        repo_id: Model repo id to inspect.
-        gpu_type: Target GPU type, used for TP/profile selection.
-
-    Returns:
-        A :class:`DetectedConfig`, or ``None`` when the HF metadata cannot be
-        fetched.
-    """
-    log.info("[%s] fetching HF metadata", repo_id)
-    try:
-        info = hf.model_info(repo_id)
-        config = hf.model_config(repo_id)
-    except Exception as e:
-        log.error("[%s] HF fetch failed: %s", repo_id, e)
-        return None
-
-    arch = (config.get("architectures") or ["unknown"])[0]
-
-    # Refuse non-generative repos even with explicit --model: sglang/vllm won't
-    # start a server for them and the task would just burn a sandbox slot.
-    if not is_generative_arch(arch):
-        log.error(
-            "[%s] arch=%s is not a generative LM "
-            "(expected ForCausalLM / ForConditionalGeneration / LMHeadModel / ForSeq2SeqLM "
-            "suffix). Skipping — pass an actual causal-LM repo, or override "
-            "with --manual --framework vllm if you really want to try.",
-            repo_id,
-            arch,
-        )
-        return None
-
-    framework = detect_framework(config)
-    precision = detect_precision(config)
-    params_b = detect_param_count(info, config)
-    max_context_tokens = detect_max_context_tokens(config)
-    mpe_raw = config.get("max_position_embeddings")
-    max_position_embeddings = int(mpe_raw) if isinstance(mpe_raw, (int, float)) and mpe_raw > 0 else 0
-    tp = detect_tp(params_b, precision, gpu_type)
-    conc = detect_concurrency(tp, framework)
-    image = detect_image(framework, repo_id)
-
-    cfg = DetectedConfig(
-        arch=arch,
-        framework=framework,
-        precision=precision,
-        tp=tp,
-        concurrency=conc,
-        image=image,
-        params_b=params_b,
-        max_context_tokens=max_context_tokens,
-        max_position_embeddings=max_position_embeddings,
-        raw_config=config,
-    )
-    log.info(
-        "[%s] arch=%s params=%.1fB context=%d framework=%s precision=%s gpu=%s tp=%d conc=%d",
-        repo_id,
-        arch,
-        params_b,
-        max_context_tokens,
-        framework,
-        precision,
-        canonical_gpu_type(gpu_type),
-        tp,
-        conc,
-    )
-    return cfg
-
-
-# ── SaFE client ─────────────────────────────────────────────────────────────────
-
-
-class SafeOptimizeClient:
-    """Thin wrapper for SaFE playground/optimization endpoints.
-
-    Reuses the same bearer token as the rest of Hyperloom CI. The API contract
-    here mirrors SaFE/scripts/optimize_submit.py (2026-05-06), in particular
-    the ``target.volume`` field added to /api/v1/playground/models.
-    """
-
-    def __init__(
-        self,
-        base_url: str,
-        token: str,
-        register_workspace: str,
-        submit_workspace: str,
-        volume: str,
-        timeout: int = 30,
-        submit_workspaces_pool: list[str] | None = None,
-    ):
-        """Initialise the SaFE client and its authenticated HTTP session.
-
-        Args:
-            base_url (str): SaFE base URL (trailing slash trimmed).
-            token (str): Bearer token for the Authorization header.
-            register_workspace (str): Workspace where models are registered and
-                downloaded (must allow RW writes to ``volume``).
-            submit_workspace (str): Workspace where optimization tasks run (must
-                allow the Sandbox scope).
-            volume (str): Wekafs volume mounted RW in ``register_workspace``.
-            timeout (int): Per-request timeout in seconds.
-            submit_workspaces_pool (list[str] | None): Optional round-robin pool
-                of submit workspaces; when set, each submit cycles through it.
-        """
-        self.base_url = base_url.rstrip("/")
-        # Where the model registers + downloads (needs RW to the volume).
-        self.register_workspace = register_workspace
-        # Where the task is created (needs Sandbox scope); may equal register.
-        self.submit_workspace = submit_workspace
-        # Optional round-robin pool: each submit_task picks the next workspace,
-        # letting a batch span multiple workspaces without manual splitting.
-        self.submit_workspaces_pool = [w.strip() for w in (submit_workspaces_pool or []) if w and w.strip()] or None
-        self._submit_ws_counter = 0
-        self.volume = volume
-        self.timeout = timeout
-        self._sess = requests.Session()
-        self._sess.headers.update(
-            {
-                "Authorization": f"Bearer {token}",
-                "Content-Type": "application/json",
-            }
-        )
-        # Honor CA bundle env so corp proxies don't break HTTPS.
-        self._sess.verify = os.environ.get("SSL_CERT_FILE", os.environ.get("REQUESTS_CA_BUNDLE", True))
-
-    def _request(self, method: str, path: str, body: dict | None = None, timeout: float | None = None) -> dict:
-        """Send an HTTP request to the SaFE API and return the parsed JSON.
-
-        Args:
-            method: HTTP method (e.g. ``GET``, ``POST``).
-            path: API path appended to ``base_url``.
-            body: Optional JSON request body.
-            timeout: Optional per-request timeout; falls back to the client
-                default.
-
-        Returns:
-            The decoded JSON response, or an empty dict when there is no body.
-
-        Raises:
-            RuntimeError: If the response status code is ``>= 400``.
-        """
-        url = f"{self.base_url}/{path.lstrip('/')}"
-        resp = self._sess.request(method, url, json=body, timeout=timeout or self.timeout)
-        if resp.status_code >= 400:
-            raise RuntimeError(f"{method} {path} -> HTTP {resp.status_code}: {resp.text[:300]}")
-        return resp.json() if resp.content else {}
-
-    def find_model(self, repo_id: str) -> dict | None:
-        """Look up an existing SaFE Model by HF source URL, scoped to
-        register_workspace (where the canonical Model CR + LocalPaths live).
-
-        Args:
-            repo_id (str): HuggingFace repo id to match by source URL.
-
-        Returns:
-            dict | None: The matching SaFE model record, or ``None`` when none
-            matches or the listing fails.
-        """
-        hf_url = f"https://huggingface.co/{repo_id}".rstrip("/")
-        from urllib.parse import quote
-
-        try:
-            data = self._request(
-                "GET",
-                f"api/v1/playground/models?limit=200&workspace={quote(self.register_workspace)}",
-            )
-        except Exception as e:
-            log.warning("list models failed: %s", e)
-            return None
-        for m in data.get("items", []):
-            if (m.get("sourceURL") or "").rstrip("/") == hf_url:
-                return m
-        return None
-
-    def register_model(
-        self,
-        repo_id: str,
-        hf_token: str = "",
-        local_path: str = "",
-    ) -> str:
-        """Register a model record with SaFE so submit_task has a model_id.
-
-        local_path set → accessMode=local_path: SaFE skips its Download Job
-        (phase=Ready immediately) since files are already on disk (prewarm path).
-        local_path empty → accessMode=local: SaFE downloads from HF (slow
-        fallback when prewarm can't run).
-
-        Args:
-            repo_id (str): HuggingFace repo id to register.
-            hf_token (str): Optional HF token for gated downloads (local mode).
-            local_path (str): On-disk model path; when set, registers in
-                ``local_path`` mode and skips SaFE's Download Job.
-
-        Returns:
-            str: The registered SaFE model id (empty string when absent).
-        """
-        if local_path:
-            # local_path mode bypasses SaFE's HF metadata fetch, so we MUST
-            # provide displayName. SaFE feeds it into GenerateName → K8s
-            # metadata.name, which must satisfy RFC 1123 (lowercase [a-z0-9-.],
-            # 1-63 chars); sanitize here since the backend doesn't.
-
-            raw = repo_id.split("/")[-1] or repo_id
-            cleaned = re.sub(r"[^a-z0-9.-]+", "-", raw.lower()).strip(".-") or "model"
-            # Trim to 50 to leave headroom for GenerateName's -xxxxx suffix (max 63).
-            display_name = cleaned[:50].rstrip(".-") or "model"
-            body = {
-                "displayName": display_name,
-                "source": {
-                    "url": repo_id,
-                    "accessMode": "local_path",
-                    "localPath": local_path,
-                },
-                "workspace": self.register_workspace,
-            }
-            log.info(
-                "[%s] register (local_path mode): workspace=%s displayName=%s localPath=%s",
-                repo_id,
-                self.register_workspace,
-                display_name,
-                local_path,
-            )
-        else:
-            body = {
-                "source": {
-                    "url": repo_id,
-                    "accessMode": "local",
-                    **({"token": hf_token} if hf_token else {}),
-                },
-                "workspace": self.register_workspace,
-                "target": {"volume": self.volume},
-            }
-            log.info(
-                "[%s] register (local mode — SaFE will download): workspace=%s volume=%s",
-                repo_id,
-                self.register_workspace,
-                self.volume,
-            )
-        result = self._request("POST", "api/v1/playground/models", body)
-        return result.get("id", "")
-
-    def wait_ready(
-        self,
-        model_id: str,
-        timeout_min: int = 480,
-        poll_s: int = 30,
-    ) -> bool:
-        """Poll a SaFE model until it reaches the Ready phase.
-
-        Args:
-            model_id (str): SaFE model id to poll.
-            timeout_min (int): Maximum minutes to wait before giving up.
-            poll_s (int): Seconds between polls.
-
-        Returns:
-            bool: True once the model is Ready; False if it reaches Failed or
-                the timeout elapses.
-        """
-        log.info("waiting for model %s to be Ready (timeout=%dm)", model_id, timeout_min)
-        deadline = time.time() + timeout_min * 60
-        last_phase = ""
-        while time.time() < deadline:
-            try:
-                m = self._request("GET", f"api/v1/playground/models/{model_id}")
-                phase = m.get("phase", "")
-                if phase != last_phase:
-                    log.info("model %s phase: %s", model_id, phase or "(empty)")
-                    last_phase = phase
-                if phase == "Ready":
-                    return True
-                if phase == "Failed":
-                    log.error("model %s Failed: %s", model_id, m.get("message", ""))
-                    return False
-            except Exception as e:
-                log.debug("phase poll error (will retry): %s", e)
-            time.sleep(poll_s)
-        log.error("model %s wait timed out after %dm", model_id, timeout_min)
-        return False
-
-    def submit_task(
-        self,
-        model_id: str,
-        display_name: str,
-        framework: str,
-        precision: str,
-        tp: int,
-        concurrency: int,
-        isl: int,
-        osl: int,
-        image: str | None,
-        mode: str = "local",
-        gpu_type: str | None = None,
-        inferencex_path: str | None = None,
-        oob_path: str | None = None,
-        tracelens_root: str | None = None,
-        prompt_prefix: str | None = None,
-        prompt_suffix: str | None = None,
-        kernel_backends: list[str] | None = None,
-        max_hours: float | None = None,
-        target_gain: float | None = None,
-        results_path: str | None = None,
-        env: dict | None = None,
-    ) -> dict:
-        """Submit an optimization task to SaFE and return the API response.
-
-        Builds the task request body from the model and benchmark parameters,
-        choosing a target workspace (single or round-robin across the pool).
-
-        Args:
-            model_id: Registered SaFE model id.
-            display_name: Human-readable task name.
-            framework: Serving framework (``vllm`` / ``sglang``).
-            precision: Model precision (e.g. ``BF16``, ``FP8``).
-            tp: Tensor-parallel size.
-            concurrency: Benchmark concurrency level.
-            isl: Input sequence length.
-            osl: Output sequence length.
-            image: Server image to run, or ``None`` for the framework default.
-            mode: Submission mode (e.g. ``local``).
-            gpu_type: Target GPU type.
-            inferencex_path: Optional InferenceX checkout path.
-            oob_path: Optional out-of-box baseline path.
-            tracelens_root: Optional TraceLens root path.
-            prompt_prefix: Optional prompt prefix override.
-            prompt_suffix: Optional prompt suffix override.
-            kernel_backends: Optional list of kernel backends to enable.
-            max_hours: Optional wall-clock budget for the task.
-            target_gain: Optional target performance gain.
-            results_path: Optional path where results should be written.
-
-        Returns:
-            The decoded API response for the submitted task.
-        """
-        # Pick the workspace: single submit_workspace, or round-robin across the
-        # pool. Counter is per-instance, not thread-safe — fine since submit_task
-        # runs serially (only wait_and_collect is parallel, after submit returns).
-        if self.submit_workspaces_pool:
-            chosen_ws = self.submit_workspaces_pool[self._submit_ws_counter % len(self.submit_workspaces_pool)]
-            self._submit_ws_counter += 1
-            log.info(
-                "[submit] round-robin chose workspace=%s (pool=%s, idx=%d)",
-                chosen_ws,
-                ",".join(self.submit_workspaces_pool),
-                self._submit_ws_counter - 1,
-            )
-        else:
-            chosen_ws = self.submit_workspace
-        body = {
-            "displayName": display_name,
-            "modelId": model_id,
-            "workspace": chosen_ws,
-            "mode": mode,
-            "framework": framework,
-            "precision": precision,
-            "tp": tp,
-            "ep": 1,
-            "isl": isl,
-            "osl": osl,
-            "concurrency": concurrency,
-            "kernelBackends": list(kernel_backends or DEFAULT_KERNEL_BACKENDS),
-        }
-        if max_hours and max_hours > 0:
-            body["maxHours"] = max_hours
-        if target_gain and target_gain > 0:
-            body["targetGain"] = target_gain
-        if results_path:
-            body["resultsPath"] = results_path
-        if image:
-            body["image"] = image
-        # Override SaFE's wrong-for-core42 MI355X default (see DEFAULT_GPU_TYPE).
-        if gpu_type:
-            body["gpuType"] = gpu_type
-        # Always send inferencexPath (even empty) to suppress SaFE's Zod default
-        # "/hyperloom/InferenceX"; empty lets install.sh clone a writable copy.
-        body["inferencexPath"] = inferencex_path or ""
-        if oob_path:
-            body["oobPath"] = oob_path
-        if tracelens_root:
-            body["tracelensRoot"] = tracelens_root
-        # Long-run tuning: when the optimizer budget exceeds 24h, ask the agent
-        # to enable phase interleaving + cyclic phases so the extra wall-clock is
-        # spent re-tuning in cycles instead of stalling after the first pass.
-        # Injected as shell exports (same prelude mechanism as HYPERLOOM_SOURCE_*).
-        if max_hours and max_hours > 24:
-            longrun_hint = (
-                "LONG-RUN TUNING (optimizer budget > 24h) — also run these exports "
-                "in the shell prelude before launching the optimizer:\n"
-                "  export INFERENCE_OPTIMIZER_PHASE_INTERLEAVE=1\n"
-                "  export INFERENCE_OPTIMIZER_CYCLIC_PHASES=1\n"
-                "  export INFERENCE_OPTIMIZER_CYCLE_HOURS=16\n\n"
-            )
-            prompt_prefix = longrun_hint + (prompt_prefix or "")
-        # Optional prefix/suffix forwarded to BuildHyperloomPrompt on the SaFE side.
-        if prompt_prefix:
-            body["promptPrefix"] = prompt_prefix
-        if prompt_suffix:
-            body["promptSuffix"] = prompt_suffix
-        # Optional session-scoped env forwarded to SaFE (body.env). SaFE relays
-        # it to Claw as session_env, injected into the sandbox so the
-        # inference_optimizer process sees it (e.g. CLAUDE_MODEL override).
-        # Logged so each CI job records exactly which session env it submitted,
-        # making "did the env reach Claw" verifiable from the job log alone.
-        if env:
-            body["env"] = env
-            log.info("[%s] session env forwarded to SaFE: %s", display_name, env)
-        attempts = 8
-        # Captured before the first POST so the dedup lookup only matches a task
-        # this call created (not an unrelated older one for the same model).
-        submit_started_at = time.time()
-        for attempt in range(1, attempts + 1):
-            try:
-                # The submit POST can be slow when the core42 apiserver is
-                # under load from many parallel daily jobs. Give it a generous
-                # read timeout so a busy-but-alive backend doesn't trip the
-                # default 30s and get misreported as a hard submit failure.
-                return self._request("POST", "api/v1/optimization/tasks", body, timeout=120)
-            except Exception as e:
-                msg = str(e)
-                low = msg.lower()
-                transient = (
-                    "HTTP 500" in msg
-                    or "HTTP 502" in msg
-                    or "HTTP 503" in msg
-                    or "HTTP 504" in msg
-                    or "timed out" in low  # requests ReadTimeout/ConnectTimeout
-                    or "timeout" in low
-                    or "connection" in low  # ConnectionError / HTTPSConnectionPool
-                )
-                if not transient:
-                    raise
-                # The POST is NOT idempotent: a slow/dropped response can hide a
-                # task the backend actually created. Blindly retrying then spawns
-                # a SECOND Claw session for the same model (the duplicate /
-                # abandoned-PRELUDE dirs). Before retrying, look the task up by
-                # modelId+workspace; if it already exists, reuse it.
-                existing = self._find_recent_submitted_task(model_id, chosen_ws, submit_started_at)
-                if existing and existing.get("id"):
-                    log.warning(
-                        "[submit] transient submit failure but a task for modelId=%s "
-                        "already exists server-side (id=%s) — reusing it instead of "
-                        "re-POSTing to avoid a duplicate session: %s",
-                        model_id,
-                        existing.get("id"),
-                        msg,
-                    )
-                    return existing
-                if attempt >= attempts:
-                    raise
-                delay = random.uniform(10, 60)
-                log.warning(
-                    "[submit] transient SaFE/Claw submit failure (attempt %d/%d, workspace=%s); retrying in %.1fs: %s",
-                    attempt,
-                    attempts,
-                    chosen_ws,
-                    delay,
-                    msg,
-                )
-                time.sleep(delay)
-        raise RuntimeError("unreachable submit retry loop exit")
-
-    def _find_recent_submitted_task(self, model_id: str, workspace: str, since_ts: float) -> dict | None:
-        """Find a task this submit call may have created before its POST failed.
-
-        A transient submit failure (slow/timeout/dropped response) can hide a
-        task the backend actually created. Look it up by ``modelId`` +
-        ``workspace`` so the retry path can reuse it instead of creating a
-        duplicate Claw session.
-
-        Args:
-            model_id (str): Registered SaFE model id used in the submit body.
-            workspace (str): Submit workspace used in the submit body.
-            since_ts (float): Unix time captured just before the first POST;
-                only tasks created at/after this (minus clock-skew slack) match.
-
-        Returns:
-            dict | None: The most recently-created matching task, or ``None``.
-        """
-        from urllib.parse import quote
-
-        try:
-            data = self._request(
-                "GET",
-                f"api/v1/optimization/tasks?modelId={quote(model_id)}"
-                f"&workspace={quote(workspace)}&limit=20",
-            )
-        except Exception as e:  # noqa: BLE001
-            log.debug("[submit] dedup lookup failed (will fall back to retry): %s", e)
-            return None
-        items = data.get("items") if isinstance(data, dict) else None
-        if not isinstance(items, list):
-            return None
-        floor = since_ts - 120.0  # 2-min slack for client/server clock skew
-        best = None
-        best_ct = floor
-        for it in items:
-            if not isinstance(it, dict) or (it.get("modelId") or "") != model_id:
-                continue
-            created = str(it.get("createdAt") or "")
-            try:
-                ct = datetime.fromisoformat(created.replace("Z", "+00:00")).timestamp()
-            except Exception:  # noqa: BLE001
-                continue
-            if ct >= best_ct:
-                best_ct = ct
-                best = it
-        return best
-
-    # ── Task lifecycle ──
-
-    # Lifecycle states from SaFE types.go OptimizationTaskStatus.
-    TERMINAL_TASK_STATUSES = {"Succeeded", "Failed", "Interrupted"}
-
-    def get_task(self, task_id: str) -> dict:
-        """Fetch the current state of an optimization task.
-
-        Args:
-            task_id (str): SaFE optimization task id.
-
-        Returns:
-            dict: The task record JSON.
-        """
-        return self._request("GET", f"api/v1/optimization/tasks/{task_id}")
-
-    def wait_task_done(
-        self,
-        task_id: str,
-        timeout_min: int = 480,
-        poll_s: int = 60,
-    ) -> tuple[str, dict]:
-        """Wait until the task reaches a terminal status. Returns (status, last_task).
-
-        Prefer the Claw SSE stream (SaFE's task status lags Claw by minutes), and
-        fall back to SaFE polling when no clawSessionId exists or SSE fails.
-        Returns ('Timeout', {}) if neither sees a terminal status by the deadline.
-
-        Args:
-            task_id (str): SaFE optimization task id to wait on.
-            timeout_min (int): Maximum minutes to wait for a terminal status.
-            poll_s (int): Seconds between SaFE status polls.
-
-        Returns:
-            tuple[str, dict]: ``(status, last_task)`` — the terminal status (or
-            ``"Timeout"``) and the last task record observed.
-        """
-        log.info("[task %s] waiting for completion (timeout=%dm, poll=%ds)", task_id, timeout_min, poll_s)
-        deadline = time.time() + timeout_min * 60
-
-        # Wait briefly (cap 60s) for clawSessionId to materialize, else fall
-        # through to polling.
-        sid = None
-        for _ in range(12):
-            sid = self._claw_session_id_for(task_id)
-            if sid:
-                break
-            time.sleep(5)
-
-        sse_used = False
-        sf_status = ""
-        # On idle_timeout/stream_error the Claw SSE merely went quiet (a long tool
-        # call, or the agent paused between phases). SaFE often *prematurely* marks
-        # such a task Failed/Interrupted ("optimization report not found") even
-        # though it is still well within budget and may resume to finalize (and
-        # write session_breakdown.json). Do NOT accept an idle-induced terminal as
-        # final: resubscribe and keep waiting, up to a bounded number of
-        # consecutive idle re-entries, before concluding. Only a real Stopped
-        # (sandbox exit), a Succeeded, the deadline, or exhausted retries end it.
-        idle_retries = 0
-        # Default 0: a single 1h idle window (idle_grace_s) is the whole budget;
-        # no resubscribe. Raise SAFE_OPTIMIZE_SSE_IDLE_RETRIES to re-enable retries.
-        max_idle_retries = int(_env_float("SAFE_OPTIMIZE_SSE_IDLE_RETRIES", 0.0))
-        while sid and time.time() < deadline:
-            sse_used = True
-            log.info("[task %s] using SSE on clawSessionId=%s", task_id, sid)
-            sse_reason = self._sse_wait_until_done(sid, deadline)
-            log.info("[task %s] SSE finished: reason=%s", task_id, sse_reason)
-            try:
-                last_task = self.get_task(task_id)
-            except Exception:
-                last_task = {}
-            sf_status = last_task.get("status", "") if last_task else ""
-            # A real success is always final.
-            if sf_status == "Succeeded":
-                return sf_status, last_task
-            # Stopped = sandbox pod exited (real end-of-task). SaFE's controller
-            # lags shutdown by 10-180s, so short-poll up to 5min for its verdict.
-            if sse_reason == "Stopped":
-                log.info("[task %s] sandbox stopped — short-polling SaFE for terminal status (up to 5min)", task_id)
-                for _ in range(30):
-                    time.sleep(10)
-                    if time.time() > deadline:
-                        break
-                    try:
-                        last_task = self.get_task(task_id)
-                    except Exception:
-                        continue
-                    sf_status = last_task.get("status", "") if last_task else ""
-                    if sf_status in self.TERMINAL_TASK_STATUSES:
-                        log.info("[task %s] SaFE settled on %s after sandbox stop", task_id, sf_status)
-                        return sf_status, last_task
-                # Sandbox gone but SaFE hasn't settled — treat as Succeeded so
-                # collect_artifacts can still read whatever the agent wrote.
-                log.info(
-                    "[task %s] SaFE never settled within 5min after "
-                    "sandbox stop — returning Succeeded (collect "
-                    "step will read ci_metrics.json directly)",
-                    task_id,
-                )
-                return "Succeeded", last_task
-            if sse_reason == "deadline":
-                return "Timeout", last_task
-            # idle_timeout / stream_error: inconclusive. Resubscribe and keep
-            # waiting (bounded) instead of accepting an idle-induced terminal.
-            if sse_reason in ("idle_timeout", "stream_error") and idle_retries < max_idle_retries and time.time() < deadline:
-                idle_retries += 1
-                log.info(
-                    "[task %s] SSE %s within budget (sf_status=%s) — resubscribing "
-                    "(idle retry %d/%d) instead of concluding on a non-Stopped status",
-                    task_id,
-                    sse_reason,
-                    sf_status or "?",
-                    idle_retries,
-                    max_idle_retries,
-                )
-                time.sleep(min(poll_s, 60))
-                continue
-            # Retries exhausted (or another reason): accept a terminal SaFE
-            # verdict if one exists, else fall through to SaFE polling.
-            if sf_status in self.TERMINAL_TASK_STATUSES:
-                return sf_status, last_task
-            log.info(
-                "[task %s] SSE inconclusive (reason=%s, sf_status=%s) — "
-                "falling back to SaFE polling for terminal status",
-                task_id,
-                sse_reason,
-                sf_status or "?",
-            )
-            break
-
-        # Fallback / continuation: SaFE optimization-API polling.
-        if not sse_used:
-            log.info("[task %s] no clawSessionId yet — using SaFE polling", task_id)
-        last_status = ""
-        last_phase = -1
-        last_task: dict = {}
-        while time.time() < deadline:
-            try:
-                t = self.get_task(task_id)
-                last_task = t
-                status = t.get("status", "")
-                phase = t.get("currentPhase", -1)
-                if status != last_status or phase != last_phase:
-                    log.info(
-                        "[task %s] status=%s phase=%s message=%s",
-                        task_id,
-                        status or "?",
-                        phase,
-                        (t.get("message") or "")[:120],
-                    )
-                    last_status, last_phase = status, phase
-                if status in self.TERMINAL_TASK_STATUSES:
-                    return status, t
-            except Exception as e:
-                log.debug("[task %s] poll error (will retry): %s", task_id, e)
-            time.sleep(poll_s)
-        log.warning("[task %s] wait timed out after %dm", task_id, timeout_min)
-        return "Timeout", last_task
-
-    def _sse_wait_until_done(self, session_id: str, deadline: float) -> str:
-        """Subscribe to the Claw session SSE stream, return when the agent ends.
-
-        IMPORTANT: do NOT return on ResultMessage — it fires at the end of EVERY
-        agent turn (dozens over 1-3h), so the first turn was being mistaken for
-        completion. The only reliable signal is sandboxStatus
-        phase=Stopped/Terminated/Failed (sandbox pod actually exits).
-
-        Returns: "Stopped" | "idle_timeout" (no events > idle_grace_s, default
-        1h) | "deadline" (per-task wall clock) | "stream_error" (caller falls
-        back).
-        """
-        url = f"{self.base_url}/claw-api/v1/chat/sessions/{session_id}/messages"
-        last_evt = time.time()
-        # Idle grace: how long the SSE may be silent (keepalive-only) before we
-        # treat it as idle. Default 1h to cover slow model download + inference
-        # server startup, which emit no agent events. Configurable via env.
-        idle_grace_s = int(_env_float("SAFE_OPTIMIZE_SSE_IDLE_GRACE_S", 3600.0))
-        try:
-            with self._sess.get(url, stream=True, timeout=(10, 60)) as r:
-                if not r.ok:
-                    log.warning("SSE stream HTTP %d for session %s", r.status_code, session_id[:8])
-                    return "stream_error"
-                current_event = None
-                for raw in r.iter_lines(decode_unicode=True):
-                    now = time.time()
-                    if now > deadline:
-                        return "deadline"
-                    if raw is None:
-                        continue
-                    if raw == "" or raw.startswith(":"):
-                        # Blank-line separator or `: keepalive` heartbeat.
-                        if now - last_evt > idle_grace_s:
-                            return "idle_timeout"
-                        continue
-                    if raw.startswith("id:"):
-                        continue
-                    if raw.startswith("event:"):
-                        current_event = raw[6:].strip()
-                        continue
-                    if not raw.startswith("data:"):
-                        continue
-                    payload = raw[5:].strip()
-                    try:
-                        d = json.loads(payload)
-                    except Exception:
-                        continue
-                    last_evt = now
-                    et = d.get("type") or current_event
-                    # ResultMessage is intentionally not a return signal (see
-                    # docstring); it still refreshes last_evt above.
-                    if et == "sandboxStatus":
-                        ph = (d.get("phase") or "").lower()
-                        if ph in ("stopped", "terminated", "failed"):
-                            return "Stopped"
-        except requests.exceptions.RequestException as e:
-            log.warning("SSE stream error for session %s: %s", session_id[:8], type(e).__name__)
-            return "stream_error"
-        except Exception as e:
-            log.warning("SSE stream unexpected error for session %s: %s", session_id[:8], e)
-            return "stream_error"
-        # Stream closed cleanly without a Stopped signal — treat as idle.
-        return "idle_timeout"
-
-    def _claw_session_id_for(self, task_id: str) -> str | None:
-        """Resolve clawSessionId for a task (per-instance cached). None when
-        SaFE has no session attached (e.g. task failed before session creation).
-
-        Args:
-            task_id (str): SaFE optimization task id.
-
-        Returns:
-            str | None: The Claw session id, or ``None`` when none is attached
-            or the lookup fails.
-        """
-        if not hasattr(self, "_claw_session_cache"):
-            self._claw_session_cache = {}
-        if task_id in self._claw_session_cache:
-            return self._claw_session_cache[task_id]
-        try:
-            t = self.get_task(task_id)
-        except Exception as e:
-            log.warning("[task %s] get_task failed while resolving clawSessionId: %s", task_id, e)
-            return None
-        sid = (t.get("clawSessionId") or "").strip()
-        self._claw_session_cache[task_id] = sid or None
-        return sid or None
-
-    def list_artifacts(self, task_id: str) -> list[dict]:
-        """List task artifacts via the SaFE standard endpoint.
-
-        Returns items shaped {path, size, lastModified, downloadPath}.
-        downloadPath is server-relative; download_artifact() uses it when present,
-        else builds the /artifacts/download?path= URL from path.
-
-        Args:
-            task_id (str): SaFE optimization task id.
-
-        Returns:
-            list[dict]: Artifact item dicts, or ``[]`` on error / unexpected
-            shape.
-        """
-        try:
-            data = self._request("GET", f"api/v1/optimization/tasks/{task_id}/artifacts")
-        except Exception as e:
-            log.warning("[task %s] list_artifacts failed: %s", task_id, e)
-            return []
-        items = (data.get("items") or data.get("data")) if isinstance(data, dict) else data
-        if isinstance(items, dict) and isinstance(items.get("items"), list):
-            items = items["items"]
-        if not isinstance(items, list):
-            return []
-        return items
-
-    def download_artifact(self, task_id: str, path_or_item: "str | dict") -> bytes:
-        """Download a single task artifact. Accepts a string path or a
-        list_artifacts item dict; prefers the item's downloadPath when present.
-
-        Args:
-            task_id (str): SaFE optimization task id.
-            path_or_item (str | dict): Artifact path string or a
-                :meth:`list_artifacts` item dict.
-
-        Returns:
-            bytes: The downloaded artifact content.
-
-        Raises:
-            RuntimeError: If the download response status is ``>= 400``.
-        """
-        if isinstance(path_or_item, dict):
-            download_path = (path_or_item.get("downloadPath") or "").strip()
-            if download_path:
-                url = f"{self.base_url}/{download_path.lstrip('/')}"
-            else:
-                path = path_or_item.get("path", "")
-                encoded = requests.utils.quote(path, safe="")
-                url = f"{self.base_url}/api/v1/optimization/tasks/{task_id}/artifacts/download?path={encoded}"
-        else:
-            encoded = requests.utils.quote(path_or_item, safe="")
-            url = f"{self.base_url}/api/v1/optimization/tasks/{task_id}/artifacts/download?path={encoded}"
-        resp = self._sess.get(url, timeout=120)
-        if resp.status_code >= 400:
-            raise RuntimeError(f"SaFE artifact download -> HTTP {resp.status_code}: {resp.text[:200]}")
-        return resp.content
-
-    def download_artifact_to(
-        self,
-        task_id: str,
-        path_or_item: "str | dict",
-        local_path: str,
-    ) -> int:
-        """Download an artifact and write it to a local file.
-
-        Creates parent directories as needed.
-
-        Args:
-            task_id (str): SaFE optimization task id.
-            path_or_item (str | dict): Artifact path string or item dict.
-            local_path (str): Destination file path.
-
-        Returns:
-            int: Number of bytes written.
-        """
-        data = self.download_artifact(task_id, path_or_item)
-        os.makedirs(os.path.dirname(local_path) or ".", exist_ok=True)
-        with open(local_path, "wb") as f:
-            f.write(data)
-        return len(data)
-
-
-# ── Per-model record ────────────────────────────────────────────────────────────
-
-
-@dataclass
-class SubmissionRecord:
-    """Per-model record tracking submission, completion, and CI delivery.
-
-    Accumulates state across the submit → wait → collect pipeline and is
-    serialised into the submission manifest.
-
-    Attributes:
-        model (str): HuggingFace repo id for this entry.
-        status (str): Local submit stage (submitted/dry-run/skipped/failed).
-        task_id (str | None): SaFE optimization task id once submitted.
-        claw_session_id (str | None): Claw session UUID SaFE created.
-        display_name (str | None): Task display name.
-        model_path (str | None): Resolved model path reported by SaFE.
-        safe_user_id (str | None): SaFE user id owning the task.
-        safe_started_at (str | None): SaFE task start timestamp.
-        safe_finished_at (str | None): SaFE task finish timestamp.
-        detected (dict | None): Auto-detected config as a dict.
-        overrides (dict): User/CLI overrides applied.
-        pool (dict): Production-pool audit metadata.
-        error (str | None): Error message when submit failed/skipped.
-        category (str | None): Coarse model shape (moe/dense/"").
-        sandbox_duration_seconds (float | None): SaFE wallclock duration.
-        final_status (str | None): SaFE terminal status.
-        final_phase (int | None): Current phase at the terminal moment.
-        final_message (str | None): SaFE task message.
-        ci_status (str | None): CI delivery status (separate from SaFE status).
-        ci_success (bool): Whether usable artifacts were delivered.
-        delivery_reason (str | None): Explanation for the CI delivery status.
-        artifacts_dir (str | None): Local directory artifacts landed in.
-        artifact_count (int): Number of collected artifacts.
-        artifact_files (list[str]): Collected artifact file paths.
-        artifact_sources (list[dict]): Provenance entries per artifact.
-    """
-
-    model: str
-    status: str = "pending"  # local stage: submitted/dry-run/skipped/failed
-    task_id: str | None = None
-    # Claw session UUID SaFE creates at submit; used to correlate ci_metrics.json
-    # under /wekafs/users/<uid>/<session>/ with the task (set in wait_and_collect_one).
-    claw_session_id: str | None = None
-    display_name: str | None = None
-    model_path: str | None = None
-    safe_user_id: str | None = None
-    safe_started_at: str | None = None
-    safe_finished_at: str | None = None
-    detected: dict | None = None
-    overrides: dict = field(default_factory=dict)
-    pool: dict = field(default_factory=dict)
-    error: str | None = None
-    # Audit fields so each persisted artifact is self-describing.
-    category: str | None = None  # moe / dense / "" — from detected.arch
-    sandbox_duration_seconds: float | None = None  # SaFE startedAt -> finishedAt
-    final_status: str | None = None  # SaFE: Succeeded/Failed/Interrupted/Timeout
-    final_phase: int | None = None  # currentPhase at terminal moment
-    final_message: str | None = None  # task.Message
-    # CI delivery status is separate from SaFE final_status: a SaFE timeout may
-    # still have written a useful session_breakdown worth publishing.
-    ci_status: str | None = None  # Delivered / Missing artifacts / ...
-    ci_success: bool = False
-    delivery_reason: str | None = None
-    artifacts_dir: str | None = None  # local dir where artifacts landed
-    artifact_count: int = 0
-    artifact_files: list[str] = field(default_factory=list)
-    artifact_sources: list[dict] = field(default_factory=list)
-
-
-# ── Per-model flow ──────────────────────────────────────────────────────────────
-
-
-def process_model(
-    repo_id: str,
-    hf: HuggingFaceClient,
-    safe: SafeOptimizeClient,
-    overrides: dict,
-    isl: int,
-    osl: int,
-    dry_run: bool,
-    hf_token: str,
-    manual_mode: bool,
-    mode: str,
-    gpu_type: str | None = None,
-    inferencex_path: str | None = None,
-    oob_path: str | None = None,
-    tracelens_root: str | None = None,
-    prompt_prefix: str | None = None,
-    prompt_suffix: str | None = None,
-    kernel_backends: list[str] | None = None,
-    max_hours: float | None = None,
-    target_gain: float | None = None,
-    results_path: str | None = None,
-    pool_metadata: dict | None = None,
-    env: dict | None = None,
-) -> SubmissionRecord:
-    """Run the full submit flow for one model: detect, register, submit.
-
-    Auto-detects (or uses manual overrides for) the launch config, ensures the
-    model is registered and Ready in SaFE (preferring prewarmed local_path
-    mode), then submits the optimization task. Short-circuits for dry-run and
-    for skip/failure conditions.
-
-    Args:
-        repo_id (str): HuggingFace repo id.
-        hf (HuggingFaceClient): Client for HF metadata lookups.
-        safe (SafeOptimizeClient): Client for SaFE register/submit calls.
-        overrides (dict): Field overrides (framework/precision/tp/...).
-        isl (int): Input sequence length.
-        osl (int): Output sequence length.
-        dry_run (bool): When True, plan only without registering/submitting.
-        hf_token (str): HuggingFace token for gated downloads.
-        manual_mode (bool): Skip auto-detect; requires ``framework`` override.
-        mode (str): Execution mode passed to SaFE (``local`` / ``claw``).
-        gpu_type (str | None): GPU type override for the prompt.
-        inferencex_path (str | None): InferenceX checkout path override.
-        oob_path (str | None): OOB checkout path override.
-        tracelens_root (str | None): TraceLens checkout path override.
-        prompt_prefix (str | None): Prompt prefix forwarded to SaFE.
-        prompt_suffix (str | None): Prompt suffix forwarded to SaFE.
-        kernel_backends (list[str] | None): Kernel optimization backends.
-        max_hours (float | None): Max optimization wall-clock hours.
-        target_gain (float | None): Target gain percentage.
-        results_path (str | None): Results path passed to the prompt builder.
-        pool_metadata (dict | None): Production-pool audit metadata.
-
-    Returns:
-        SubmissionRecord: The record describing the outcome of this model.
-    """
-    rec = SubmissionRecord(
-        model=repo_id,
-        overrides={k: v for k, v in overrides.items() if v is not None},
-        pool={k: v for k, v in (pool_metadata or {}).items() if v not in (None, "")},
-    )
-    gpu_type = canonical_gpu_type(gpu_type)
-    gpu_profile = normalize_gpu_profile(gpu_type, warn=False) or DEFAULT_GPU_PROFILE
-
-    detected = None if manual_mode else auto_detect(hf, repo_id, gpu_type=gpu_type)
-    if not detected and not manual_mode:
-        rec.status = "skipped"
-        rec.error = "auto-detect failed"
-        return rec
-    if manual_mode and not overrides.get("framework"):
-        rec.status = "skipped"
-        rec.error = "manual mode requires --framework"
-        return rec
-    if detected:
-        rec.detected = asdict(detected)
-        rec.category = _category_from_arch(rec.detected.get("arch", ""))
-        # Absolute small-context floor: skip models whose config.json
-        # max_position_embeddings is present and <= MIN_MAX_POSITION_EMBEDDINGS,
-        # regardless of the requested isl/osl. (A missing config.json already
-        # fails auto_detect above and is skipped there.)
-        mpe = int(rec.detected.get("max_position_embeddings") or 0)
-        if mpe and mpe <= MIN_MAX_POSITION_EMBEDDINGS:
-            rec.status = "skipped"
-            rec.error = (
-                "context_too_short: "
-                f"max_position_embeddings={mpe} <= {MIN_MAX_POSITION_EMBEDDINGS}"
-            )
-            log.warning("[%s] skipping: %s", repo_id, rec.error)
-            return rec
-        max_context = int(rec.detected.get("max_context_tokens") or 0)
-        if context_too_short(max_context, isl, osl):
-            required = isl + osl + DEFAULT_CONTEXT_RESERVE_TOKENS
-            rec.status = "skipped"
-            rec.error = (
-                "context_too_short: "
-                f"max_context_tokens={max_context} < required={required} "
-                f"(isl={isl}, osl={osl}, reserve={DEFAULT_CONTEXT_RESERVE_TOKENS})"
-            )
-            log.warning("[%s] skipping: %s", repo_id, rec.error)
-            return rec
-
-        # Shared structural compatibility pre-flight (multimodal, Gemma2,
-        # Phi3 longrope, dual-chunk attention, ModelOpt FP8, FlashInfer,
-        # missing tokenizer). Uses the real downloaded config + local model dir
-        # so doomed models are skipped before a Claw session is created.
-        compat = model_compat.unrunnable_reason(
-            detected.raw_config,
-            repo=repo_id,
-            model_dir=os.path.join(
-                os.environ.get("CI_MODELS_DIR", "/wekafs/models"),
-                repo_id.replace("/", "-")),
-            whitelist=model_compat.load_whitelist(),
-            gpu_type=gpu_type,
-        )
-        if compat:
-            reason, detail = compat
-            rec.status = "skipped"
-            rec.error = f"{reason}: {detail}"
-            log.warning(
-                "[%s] PRE-FLIGHT FILTERED — repo_id=%s rule=%s reason=%s "
-                "arch=%s params=%.1fB max_position_embeddings=%s framework=%s — "
-                "skipping: NOT submitting, no Claw session created",
-                repo_id,
-                repo_id,
-                reason,
-                detail,
-                detected.arch,
-                detected.params_b,
-                detected.max_position_embeddings or "?",
-                detected.framework,
-            )
-            return rec
-
-        # Online fallback for missing_tokenizer: the local rule above only fires
-        # once the model dir is populated with weights. Not-yet-cached repos that
-        # ship weights but omit tokenizer.* files would otherwise slip through and
-        # fail to serve. Probe the HF file listing directly (HF_TOKEN/HF_TOKEN_2)
-        # so they are skipped before a Claw session is created.
-        hf_tokens = [t for t in (os.environ.get("HF_TOKEN", ""),
-                                  os.environ.get("HF_TOKEN_2", "")) if t]
-        tok_reason = model_compat.hf_missing_tokenizer(repo_id, hf_tokens)
-        if tok_reason:
-            rec.status = "skipped"
-            rec.error = f"{tok_reason}: weights present on HF but no tokenizer files"
-            log.warning(
-                "[%s] PRE-FLIGHT FILTERED — rule=%s — skipping: NOT submitting, "
-                "no Claw session created", repo_id, tok_reason)
-            return rec
-
-    framework = overrides.get("framework") or (detected.framework if detected else "")
-    precision = overrides.get("precision") or (detected.precision if detected else "FP8")
-    tp = overrides.get("tp") or (detected.tp if detected else 1)
-    conc = overrides.get("concurrency") or (detected.concurrency if detected else 64)
-    image = overrides.get("image") or (detected.image if detected else detect_image(framework, repo_id))
-
-    log.info(
-        "[%s] => mode=%s framework=%s precision=%s tp=%d conc=%d image=%s",
-        repo_id,
-        mode,
-        framework,
-        precision,
-        tp,
-        conc,
-        image,
-    )
-
-    display_name = f"{repo_id.split('/')[-1]}-{precision.lower()}-{framework}-{gpu_profile}"
-    rec.display_name = display_name
-    rec.overrides["mode"] = mode
-    if kernel_backends:
-        rec.overrides["kernel_backends"] = kernel_backends
-    if max_hours:
-        rec.overrides["max_hours"] = max_hours
-    if target_gain:
-        rec.overrides["target_gain"] = target_gain
-    if results_path:
-        rec.overrides["results_path"] = results_path
-
-    if dry_run:
-        rec.status = "dry-run"
-        return rec
-
-    # If prewarm already populated /wekafs/models/<slug>/, use local_path mode so
-    # SaFE sets phase=Ready immediately without re-downloading over our files.
-    nfs_root = os.environ.get("NFS_ROOT", "/wekafs")
-    target_slug = repo_id.replace("/", "-")
-    target_dir = f"{nfs_root}/models/{target_slug}"
-    use_local_path = False
-    try:
-        if os.path.isdir(target_dir):
-            # Heuristic floor: any real HF repo has >=5 files (config + tokenizer
-            # + weight shard).
-            n_files = sum(1 for _ in os.scandir(target_dir))
-            if n_files >= 5:
-                use_local_path = True
-                log.info(
-                    "[%s] prewarm complete (%d files at %s) — registering "
-                    "via local_path mode (skips SaFE Download Job)",
-                    repo_id,
-                    n_files,
-                    target_dir,
-                )
-            else:
-                log.info(
-                    "[%s] %s has only %d entries — falling back to SaFE download via accessMode=local",
-                    repo_id,
-                    target_dir,
-                    n_files,
-                )
-    except OSError as e:
-        log.warning("[%s] could not probe %s: %s — falling back to SaFE download", repo_id, target_dir, e)
-
-    # Find existing SaFE record OR register fresh. Stale phase=Failed records
-    # (aborted prior download) would make wait_ready return False forever, so we
-    # re-register: SaFE issues a new model_id or resets to Pending, and the
-    # Download Job now sees prewarmed files and finishes in seconds.
-    safe_model = safe.find_model(repo_id)
-    if safe_model and safe_model.get("phase") != "Failed":
-        model_id = safe_model["id"]
-        phase = safe_model.get("phase", "")
-        log.info("[%s] found in SaFE: id=%s phase=%s", repo_id, model_id, phase)
-        if phase != "Ready" and not safe.wait_ready(model_id):
-            rec.status = "failed"
-            rec.error = "model never reached Ready"
-            return rec
-    else:
-        if safe_model:
-            log.info(
-                "[%s] existing model %s is %s — re-registering (prewarm should have populated /wekafs/models/ already)",
-                repo_id,
-                safe_model.get("id"),
-                safe_model.get("phase"),
-            )
-        try:
-            model_id = safe.register_model(
-                repo_id,
-                hf_token,
-                local_path=target_dir if use_local_path else "",
-            )
-        except Exception as e:
-            rec.status = "failed"
-            rec.error = f"register: {e}"
-            return rec
-        if not model_id:
-            rec.status = "failed"
-            rec.error = "register returned empty id"
-            return rec
-        if safe_model and model_id == safe_model.get("id"):
-            log.warning(
-                "[%s] SaFE returned the same id %s as the existing "
-                "Failed record — backend deduped by sourceURL and did "
-                "not reset phase. DELETE the record manually and rerun.",
-                repo_id,
-                model_id,
-            )
-        if not safe.wait_ready(model_id):
-            rec.status = "failed"
-            rec.error = "model never reached Ready"
-            return rec
-
-    try:
-        task_prompt_prefix = prompt_prefix
-        result = safe.submit_task(
-            model_id,
-            display_name,
-            framework,
-            precision,
-            tp,
-            conc,
-            isl,
-            osl,
-            image,
-            mode=mode,
-            gpu_type=gpu_type,
-            inferencex_path=inferencex_path,
-            oob_path=oob_path,
-            tracelens_root=tracelens_root,
-            prompt_prefix=task_prompt_prefix,
-            prompt_suffix=prompt_suffix,
-            kernel_backends=kernel_backends,
-            max_hours=max_hours,
-            target_gain=target_gain,
-            results_path=results_path,
-            env=env,
-        )
-    except Exception as e:
-        rec.status = "failed"
-        rec.error = f"submit_task: {e}"
-        return rec
-
-    rec.status = "submitted"
-    rec.task_id = result.get("id", "?")
-    log.info("[%s] OK — task_id=%s display=%s", repo_id, rec.task_id, display_name)
-    return rec
-
-
-# ── Post-submission: wait for completion + collect artifacts ───────────────────
-
-# Default artifact filter — files SaFE has the agent copy to /workspace at end of
-# Phase 10, plus variants seen in production.
-DEFAULT_ARTIFACT_PATTERNS = (
-    "optimization_report",  # optimization_report.md / *-optimization_report.md / ...
-    "ci_metrics.json",
-    # Key result preferred by claw-stats-service / V2 dashboard over ci_metrics;
-    # part of _KEY_RESULT_SUFFIXES, so missing it triggers the NFS fallback.
-    "session_breakdown.json",
-    "baseline_summary.json",
-    "sweep_results.csv",
-    "sweep_results.txt",
-    "kernel_candidates.json",
-    "kernel_results.json",
-    "run_context.env",
-    "gpu_timeline.csv",
-    "ci_summary.json",
-    "ci_report.md",
-)
-
-
-def _is_wanted_artifact(path: str, all_artifacts: bool) -> bool:
-    """Decide whether an artifact path should be downloaded.
-
-    Args:
-        path (str): The remote artifact path.
-        all_artifacts (bool): When True, accept every artifact.
-
-    Returns:
-        bool: True when ``all_artifacts`` is set or the path matches a default
-            artifact pattern.
-    """
-    if all_artifacts:
-        return True
-    p = path.lower()
-    return any(pat in p for pat in DEFAULT_ARTIFACT_PATTERNS)
-
-
-def _safe_local_path(artifacts_dir: Path, task_id: str, remote_path: str) -> Path:
-    """Map a session-relative remote path to a local file path, stripping leading
-    slashes / '..' so we never escape the artifacts dir.
-
-    Args:
-        artifacts_dir (Path): Local artifacts root.
-        task_id (str): SaFE task id used as the per-task subdirectory.
-        remote_path (str): Session-relative remote artifact path.
-
-    Returns:
-        Path: A safe local destination path under ``artifacts_dir/task_id``.
-    """
-    rel = remote_path.lstrip("/").replace("\\", "/")
-    parts = [seg for seg in rel.split("/") if seg and seg != ".." and seg != "."]
-    return artifacts_dir / task_id / Path(*parts) if parts else artifacts_dir / task_id / "artifact.bin"
-
-
-def _record_artifact_source(
-    rec: SubmissionRecord,
-    local_path: Path,
-    source_type: str,
-    *,
-    remote_path: str | None = None,
-    source_path: str | None = None,
-    session_dir: str | None = None,
-) -> None:
-    """Append a provenance entry describing where an artifact came from.
-
-    Args:
-        rec (SubmissionRecord): Record whose ``artifact_sources`` is appended.
-        local_path (Path): Local path the artifact was written to.
-        source_type (str): Origin label (e.g. ``safe_artifact_api``, ``nfs_*``).
-        remote_path (str | None): Remote artifact path, when applicable.
-        source_path (str | None): Source filesystem path, when applicable.
-        session_dir (str | None): Originating session directory, when known.
-    """
-    entry = {
-        "source_type": source_type,
-        "local_path": str(local_path).replace("\\", "/"),
-        "file_name": local_path.name,
-    }
-    if remote_path:
-        entry["remote_path"] = remote_path
-    if source_path:
-        entry["source_path"] = source_path
-    if session_dir:
-        entry["session_dir"] = session_dir
-    rec.artifact_sources.append(entry)
-
-
-def _write_artifact_sources(task_dir: Path, rec: SubmissionRecord) -> None:
-    """Write the per-task ``artifact_sources.json`` provenance file.
-
-    No-op when the record has no recorded artifact sources.
-
-    Args:
-        task_dir (Path): Per-task directory the file is written into.
-        rec (SubmissionRecord): Record carrying the artifact source entries.
-    """
-    if not rec.artifact_sources:
-        return
-    payload = {
-        "task_id": rec.task_id,
-        "model": rec.model,
-        "claw_session_id": rec.claw_session_id,
-        "artifact_dir": str(task_dir).replace("\\", "/"),
-        "files": rec.artifact_sources,
-    }
-    task_dir.mkdir(parents=True, exist_ok=True)
-    (task_dir / "artifact_sources.json").write_text(
-        json.dumps(payload, indent=2),
-        encoding="utf-8",
-    )
-
-
-# Files required for a run to count as "delivered". The NFS legacy fallback
-# scans these suffixes; a miss triggers the wekafs rescue path.
-_KEY_RESULT_SUFFIXES: tuple[str, ...] = (
-    "optimization_report.md",
-    "ci_metrics.json",
-    "session_breakdown.json",
-)
-
-
-def _norm_token(s: str) -> str:
-    """Aggressively normalise a string for fuzzy equality comparison.
-
-    Lowercases and strips dashes, underscores, dots, slashes, and spaces.
-
-    Args:
-        s (str): String to normalise.
-
-    Returns:
-        str: The normalised token.
-    """
-    return (s or "").lower().replace("-", "").replace("_", "").replace(".", "").replace("/", "").replace(" ", "")
-
-
-def _slug_token(s: str) -> str:
-    """Convert a string to a lowercase dash-separated slug.
-
-    Collapses any run of non-alphanumeric characters to a single dash and
-    trims leading/trailing dashes.
-
-    Args:
-        s (str): String to slugify.
-
-    Returns:
-        str: The slugified token.
-    """
-    out = []
-    prev_dash = False
-    for ch in (s or "").lower():
-        if ch.isalnum():
-            out.append(ch)
-            prev_dash = False
-        elif not prev_dash:
-            out.append("-")
-            prev_dash = True
-    return "".join(out).strip("-")
-
-
-def _metrics_have_positive_throughput(path: str) -> bool:
-    """Report whether a ci_metrics.json file has real, non-zero throughput.
-
-    Args:
-        path (str): Path to a ``ci_metrics.json`` file.
-
-    Returns:
-        bool: True when both baseline and optimized throughput parse to values
-            greater than zero; False on any read/parse error.
-    """
-    try:
-        with open(path, encoding="utf-8") as f:
-            data = json.load(f)
-    except Exception:
-        return False
-    baseline = data.get("baseline_throughput") or data.get("tok_per_gpu_baseline")
-    optimized = data.get("optimized_throughput") or data.get("tok_per_gpu_optimized")
-    try:
-        return float(baseline) > 0 and float(optimized) > 0
-    except (TypeError, ValueError):
-        return False
-
-
-def _json_has_any_number(value) -> bool:
-    """Recursively test whether a JSON value contains any real number.
-
-    Booleans are not counted as numbers; lists are scanned up to the first 100
-    elements.
-
-    Args:
-        value: Any JSON-decoded value.
-
-    Returns:
-        bool: True when an int/float (non-bool) is found anywhere within.
-    """
-    if isinstance(value, bool):
-        return False
-    if isinstance(value, (int, float)):
-        return True
-    if isinstance(value, dict):
-        return any(_json_has_any_number(v) for v in value.values())
-    if isinstance(value, list):
-        return any(_json_has_any_number(v) for v in value[:100])
-    return False
-
-
-def _breakdown_has_basic_data(path: Path) -> bool:
-    """True when a session_breakdown JSON carries usable audit/perf payload. The
-    delivery contract is "has structured data", not "has positive gain" (aborted
-    runs may legitimately be zero).
-
-    Args:
-        path (Path): Path to a ``session_breakdown`` JSON file.
-
-    Returns:
-        bool: True when the JSON is an object carrying a known audit/perf key or
-        any numeric value; False on read/parse error.
-    """
-    try:
-        with open(path, encoding="utf-8") as f:
-            data = json.load(f)
-    except Exception:
-        return False
-    if not isinstance(data, dict):
-        return False
-    keys = (
-        "baseline",
-        "best",
-        "final",
-        "optimized",
-        "steps",
-        "actions",
-        "phases",
-        "session",
-        "session_meta",
-        "workload",
-        "baseline_throughput",
-        "optimized_throughput",
-        "tok_per_gpu_baseline",
-        "tok_per_gpu_optimized",
-        "gain_pct",
-        "cumulative_gain_pct",
-    )
-    if any(data.get(k) not in (None, {}, [], "") for k in keys):
-        return True
-    return _json_has_any_number(data)
-
-
-def _mark_record_delivery(rec: SubmissionRecord) -> None:
-    """Set CI-level delivery status from collected artifacts.
-
-    Scans the record's artifact files/dir for a publishable
-    ``session_breakdown`` JSON and updates ``ci_success``, ``ci_status``, and
-    ``delivery_reason`` accordingly.
-
-    Args:
-        rec (SubmissionRecord): The submission record to mutate in place.
-    """
-    candidates: list[Path] = []
-    for raw in rec.artifact_files:
-        p = Path(raw)
-        if p.is_file() and p.name.startswith("session_breakdown") and p.suffix == ".json":
-            candidates.append(p)
-    if rec.artifacts_dir:
-        root = Path(rec.artifacts_dir)
-        if root.is_dir():
-            candidates.extend(p for p in root.glob("**/session_breakdown*.json") if p.is_file())
-    # De-duplicate while preserving order.
-    seen: set[str] = set()
-    unique = []
-    for p in candidates:
-        key = str(p.resolve()) if p.exists() else str(p)
-        if key not in seen:
-            seen.add(key)
-            unique.append(p)
-
-    for p in unique:
-        if _breakdown_has_basic_data(p):
-            rec.ci_success = True
-            rec.ci_status = "Delivered"
-            rec.delivery_reason = f"publishable session_breakdown: {p.name}"
-            return
-
-    if rec.artifact_count:
-        rec.ci_status = "Artifacts incomplete"
-        rec.delivery_reason = "artifacts collected but no usable session_breakdown"
-    else:
-        rec.ci_status = "Missing artifacts"
-        rec.delivery_reason = "no artifacts collected"
-
-
-def _timestamp_hint_variants(value: str) -> set[str]:
-    """Return path-matchable variants for skill session timestamps.
-
-    Args:
-        value (str): A session timestamp string (e.g. ``"20260512T010203Z"``).
-
-    Returns:
-        set[str]: Case- and separator-normalized variants for substring
-            matching against paths; empty if ``value`` is blank.
-    """
-    raw = value.strip()
-    if not raw:
-        return set()
-    compact = raw.replace("T", "").replace("t", "").replace("Z", "").replace("z", "")
-    variants = {raw, raw.lower()}
-    if compact and compact != raw:
-        variants.update({compact, compact.lower()})
-    return variants
-
-
-def _session_hints_from_artifact_items(items: list[dict]) -> set[str]:
-    """Collect session-timestamp hints from artifact item metadata.
-
-    Args:
-        items (list[dict]): Artifact item dicts with ``path``/``downloadPath``/
-            ``name`` fields.
-
-    Returns:
-        set[str]: Timestamp hint variants discovered across the items.
-    """
-    hints: set[str] = set()
-    for item in items or []:
-        if not isinstance(item, dict):
-            continue
-        text = " ".join(str(item.get(key) or "") for key in ("path", "downloadPath", "name"))
-        for match in re.findall(r"\b\d{8}T\d{6}Z\b", text, flags=re.IGNORECASE):
-            hints.update(_timestamp_hint_variants(match))
-    return hints
-
-
-def _path_has_session_hint(path: str, hints: set[str]) -> bool:
-    """Report whether a path contains any of the session timestamp hints.
-
-    Args:
-        path (str): The path to test.
-        hints (set[str]): Timestamp hint variants to look for.
-
-    Returns:
-        bool: True if any hint appears in the normalized path.
-    """
-    if not hints:
-        return False
-    norm_path = _norm_token(path)
-    return any(_norm_token(hint) in norm_path for hint in hints)
-
-
-def _parse_safe_timestamp(value: str | None) -> datetime | None:
-    """Parse a SaFE ISO-8601 timestamp into a UTC datetime.
-
-    Args:
-        value (str | None): An ISO-8601 string (``Z`` suffix accepted).
-
-    Returns:
-        datetime | None: A timezone-aware UTC datetime, or ``None`` if blank
-            or unparseable.
-    """
-    raw = (value or "").strip()
-    if not raw:
-        return None
-    try:
-        dt = datetime.fromisoformat(raw.replace("Z", "+00:00"))
-    except Exception:
-        return None
-    if dt.tzinfo is None:
-        dt = dt.replace(tzinfo=timezone.utc)
-    return dt.astimezone(timezone.utc)
-
-
-def _parse_session_timestamp(value: str) -> datetime | None:
-    """Parse a compact ``YYYYMMDDTHHMMSSZ`` session timestamp.
-
-    Args:
-        value (str): The compact session timestamp string.
-
-    Returns:
-        datetime | None: A UTC datetime, or ``None`` if blank or unparseable.
-    """
-    raw = value.strip()
-    if not raw:
-        return None
-    try:
-        return datetime.strptime(raw.upper(), "%Y%m%dT%H%M%SZ").replace(tzinfo=timezone.utc)
-    except Exception:
-        return None
-
-
-def _session_timestamp_from_path(path: str) -> str:
-    """Extract the last ``YYYYMMDDTHHMMSSZ`` timestamp found in a path.
-
-    Args:
-        path (str): The path to scan.
-
-    Returns:
-        str: The matched timestamp (upper-cased), or ``""`` if none found.
-    """
-    matches = re.findall(r"\b\d{8}T\d{6}Z\b", path, flags=re.IGNORECASE)
-    return matches[-1].upper() if matches else ""
-
-
-def _timestamp_in_task_window(timestamp: str, rec: SubmissionRecord, margin_hours: int = 2) -> bool:
-    """Check whether a session timestamp falls within the task's run window.
-
-    Args:
-        timestamp (str): A compact session timestamp string.
-        rec (SubmissionRecord): The record providing SaFE start/finish times.
-        margin_hours (int): Slack added on each side of the window.
-
-    Returns:
-        bool: True if the timestamp lies within the (padded) task window.
-    """
-    ts = _parse_session_timestamp(timestamp)
-    start = _parse_safe_timestamp(rec.safe_started_at)
-    end = _parse_safe_timestamp(rec.safe_finished_at)
-    if ts is None or start is None:
-        return False
-    if end is None:
-        end = start + timedelta(hours=24)
-    return (start - timedelta(hours=margin_hours)) <= ts <= (end + timedelta(hours=margin_hours))
-
-
-def _record_has_task_window(rec: SubmissionRecord) -> bool:
-    """Report whether a record has a usable SaFE start timestamp.
-
-    Args:
-        rec (SubmissionRecord): The submission record to inspect.
-
-    Returns:
-        bool: True if ``safe_started_at`` parses into a timestamp.
-    """
-    return _parse_safe_timestamp(rec.safe_started_at) is not None
-
-
-def _env_float(name: str, default: float) -> float:
-    """Read a float environment setting with a safe fallback.
-
-    Args:
-        name (str): Environment variable name to read.
-        default (float): Fallback value when unset or non-float.
-
-    Returns:
-        float: The parsed value, or ``default``.
-    """
-    raw = os.environ.get(name, "").strip()
-    if not raw:
-        return default
-    try:
-        return float(raw)
-    except ValueError:
-        log.warning("%s=%r is not a float; using %.1f", name, raw, default)
-        return default
-
-
-def _read_session_state(session_dir: str | Path) -> dict:
-    """Best-effort read of a session's state.json.
-
-    Args:
-        session_dir (str | Path): Session directory containing ``state.json``.
-
-    Returns:
-        dict: The parsed state dict, or an empty dict when missing/unreadable
-        or not a JSON object.
-    """
-    path = Path(session_dir) / "state.json"
-    if not path.is_file():
-        return {}
-    try:
-        data = json.loads(path.read_text(encoding="utf-8", errors="replace"))
-    except Exception:
-        return {}
-    return data if isinstance(data, dict) else {}
-
-
-def _session_has_terminal_marker(session_dir: str | Path) -> bool:
-    """True when a session has reached a state where CI should collect now.
-
-    Args:
-        session_dir (str | Path): Session directory to inspect.
-
-    Returns:
-        bool: True when a ``session_breakdown.json`` / ``complete`` marker
-        exists or ``state.json`` reports ``close_sequence_done``.
-    """
-    root = Path(session_dir)
-    if (root / "session_breakdown.json").is_file():
-        return True
-    if (root / "complete").is_file():
-        return True
-    state = _read_session_state(root)
-    if state.get("close_sequence_done") is True:
-        return True
-    return False
-
-
-def _session_activity_mtime(session_dir: str | Path) -> float:
-    """Return a bounded best-effort activity timestamp for a session.
-
-    ``state.json`` can be quiet while long Magpie subprocesses append logs or
-    traces, so include the runtime subtrees CI relies on. The file cap avoids
-    expensive walks for very large sessions.
-
-    Args:
-        session_dir (str | Path): Session directory to scan.
-
-    Returns:
-        float: The newest relevant file mtime, or ``0.0`` when nothing is found.
-    """
-    root = Path(session_dir)
-    mtimes: list[float] = []
-    for rel in (
-        "state.json",
-        "session_breakdown.json",
-        "complete",
-        "reports/final.md",
-        "reports/final.json",
-    ):
-        p = root / rel
-        try:
-            if p.exists():
-                mtimes.append(p.stat().st_mtime)
-        except OSError:
-            continue
-
-    seen = 0
-    for sub in ("optimizer_runs", "runs", "reports"):
-        base = root / sub
-        if not base.exists():
-            continue
-        for walk_root, _dirs, files in os.walk(base):
-            for name in files:
-                seen += 1
-                if seen > 5000:
-                    return max(mtimes) if mtimes else 0.0
-                if not name.endswith((".log", ".json", ".txt", ".md", ".csv", ".gz")):
-                    continue
-                try:
-                    mtimes.append((Path(walk_root) / name).stat().st_mtime)
-                except OSError:
-                    continue
-    return max(mtimes) if mtimes else 0.0
-
-
-def _find_nfs_state_session_dir(
-    rec: SubmissionRecord,
-    current_session_hints: set[str] | None = None,
-) -> str | None:
-    """Locate the current NFS session using state.json, not breakdown files.
-
-    Args:
-        rec (SubmissionRecord): Record providing user id, model names, and task
-            window for matching.
-        current_session_hints (set[str] | None): Optional session-timestamp
-            hints that, when present, must appear in the matched path.
-
-    Returns:
-        str | None: The best-matching session directory path, or ``None`` when
-        none matches.
-    """
-    nfs_root = os.environ.get("NFS_ROOT", "/wekafs")
-    users_root = Path(nfs_root) / "users"
-    if not rec.safe_user_id or not users_root.is_dir():
-        return None
-    uid_path = users_root / rec.safe_user_id
-    if not uid_path.is_dir():
-        return None
-
-    hints = set(current_session_hints or set())
-    candidates: list[tuple[int, float, str]] = []
-    for model_dir_name in _candidate_model_dir_names(rec):
-        model_dir = uid_path / model_dir_name
-        if not model_dir.is_dir():
-            continue
-        try:
-            ts_entries = sorted(os.listdir(model_dir), reverse=True)
-        except OSError:
-            continue
-        for ts_entry in ts_entries:
-            session_dir = model_dir / ts_entry
-            if not session_dir.is_dir():
-                continue
-            state_path = session_dir / "state.json"
-            if not state_path.is_file():
-                continue
-            if hints:
-                if not _path_has_session_hint(str(session_dir), hints):
-                    continue
-                score = 40
-            else:
-                ts = _session_timestamp_from_path(ts_entry)
-                if not _timestamp_in_task_window(ts, rec):
-                    continue
-                score = 30
-            state = _read_session_state(session_dir)
-            workload = state.get("workload") if isinstance(state.get("workload"), dict) else {}
-            model_field = str(state.get("model") or state.get("model_name") or workload.get("model_name") or "")
-            if model_field and not _record_model_field_matches(rec, model_field):
-                continue
-            if model_field:
-                score += 100
-            try:
-                mtime = state_path.stat().st_mtime
-            except OSError:
-                continue
-            candidates.append((score, mtime, str(session_dir)))
-
-    if not candidates:
-        return None
-    candidates.sort(reverse=True)
-    return candidates[0][2]
-
-
-def _wait_for_nfs_session_delivery(
-    rec: SubmissionRecord,
-    current_session_hints: set[str] | None = None,
-    poll_s: int = 60,
-    grace_min: float | None = None,
-    idle_min: float | None = None,
-) -> str | None:
-    """After SaFE early terminal, wait while the NFS session is still active.
-
-    Args:
-        rec (SubmissionRecord): Record used to locate the NFS session.
-        current_session_hints (set[str] | None): Optional session-timestamp
-            hints to constrain the matched session.
-        poll_s (int): Seconds between activity polls.
-        grace_min (float | None): Max minutes to wait; ``None`` resolves from
-            ``$SAFE_OPTIMIZE_NFS_LIVE_GRACE_MIN``.
-        idle_min (float | None): Idle minutes that end the wait; ``None``
-            resolves from ``$SAFE_OPTIMIZE_NFS_IDLE_GRACE_MIN``.
-
-    Returns:
-        str | None: The session directory once delivery/idle settles, or
-        ``None`` when none is found or grace waiting is disabled.
-    """
-    grace_min = _env_float("SAFE_OPTIMIZE_NFS_LIVE_GRACE_MIN", 180.0) if grace_min is None else grace_min
-    idle_min = _env_float("SAFE_OPTIMIZE_NFS_IDLE_GRACE_MIN", 20.0) if idle_min is None else idle_min
-    if grace_min <= 0 or idle_min <= 0:
-        return None
-
-    session_dir = _find_nfs_state_session_dir(rec, current_session_hints)
-    if not session_dir:
-        return None
-
-    now = time.time()
-    activity = _session_activity_mtime(session_dir)
-    if not activity:
-        return session_dir
-    if now - activity > idle_min * 60 and not _session_has_terminal_marker(session_dir):
-        log.info(
-            "[task %s] NFS session %s found but inactive for %.1fmin; collecting without grace wait",
-            rec.task_id,
-            session_dir,
-            (now - activity) / 60,
-        )
-        return session_dir
-
-    deadline = now + grace_min * 60
-    idle_deadline = activity + idle_min * 60
-    log.warning(
-        "[task %s] SaFE status=%s but NFS session still appears active: %s; "
-        "waiting up to %.1fmin (idle %.1fmin) for delivery contract files",
-        rec.task_id,
-        rec.final_status,
-        session_dir,
-        grace_min,
-        idle_min,
-    )
-    while time.time() < deadline:
-        if _session_has_terminal_marker(session_dir):
-            log.info(
-                "[task %s] NFS session reached terminal/delivery marker: %s",
-                rec.task_id,
-                session_dir,
-            )
-            return session_dir
-        latest = _session_activity_mtime(session_dir)
-        if latest > activity:
-            activity = latest
-            idle_deadline = latest + idle_min * 60
-            log.info(
-                "[task %s] NFS session still active (last activity %s)",
-                rec.task_id,
-                datetime.fromtimestamp(latest, tz=timezone.utc).isoformat(),
-            )
-        if time.time() > idle_deadline:
-            log.warning(
-                "[task %s] NFS session idle for %.1fmin without delivery marker; proceeding to collect",
-                rec.task_id,
-                idle_min,
-            )
-            return session_dir
-        time.sleep(max(1, min(poll_s, 60)))
-
-    log.warning(
-        "[task %s] NFS live-session grace wait expired after %.1fmin; proceeding to collect",
-        rec.task_id,
-        grace_min,
-    )
-    return session_dir
-
-
-def _category_from_arch(arch: str | None) -> str:
-    """Coarse model-shape classification: "moe" if arch contains "moe", else
-    "dense"; "" when unknown so downstream JSON stays "n/a".
-
-    Args:
-        arch (str | None): HF architecture class name.
-
-    Returns:
-        str: ``"moe"``, ``"dense"``, or ``""`` for empty/unknown input.
-    """
-    if not arch:
-        return ""
-    return "moe" if "moe" in arch.lower() else "dense"
-
-
-def _sandbox_duration_seconds(last_task: dict) -> float | None:
-    """SaFE-side sandbox wallclock = finishedAt - startedAt (from the task API).
-    None when either field is missing/unparseable so we don't fabricate a duration.
-
-    Args:
-        last_task (dict): SaFE task record carrying ``startedAt``/``finishedAt``.
-
-    Returns:
-        float | None: Rounded duration in seconds, or ``None`` when either
-        timestamp is missing/unparseable or the delta is negative.
-    """
-    from datetime import datetime
-
-    start = (last_task or {}).get("startedAt") or ""
-    end = (last_task or {}).get("finishedAt") or ""
-    if not start or not end:
-        return None
-    try:
-        # SaFE serializes UTC with trailing 'Z'; fromisoformat needs '+00:00'.
-        s = datetime.fromisoformat(start.replace("Z", "+00:00"))
-        e = datetime.fromisoformat(end.replace("Z", "+00:00"))
-    except Exception:
-        return None
-    delta = (e - s).total_seconds()
-    return round(delta, 1) if delta >= 0 else None
-
-
-def _find_hyperloom_commit_sha(start: Path) -> str:
-    """Resolve the Hyperloom git SHA the sandbox cloned (for audit fields).
-
-    First hit wins: (1) hyperloom_source_commit.txt written by the agent (depth
-    varies by which fallback collected it), then (2) the CI runner env
-    (HYPERLOOM_SOURCE_REF, else GITHUB_SHA).
-
-    Args:
-        start (Path): A path near the artifact, used to derive sibling
-            ``hyperloom_source_commit.txt`` candidates.
-
-    Returns:
-        str: The resolved SHA-shaped commit string, or ``""`` when none is
-        found.
-    """
-    candidates = [
-        start.parent / "hyperloom_source_commit.txt",
-        start.parent / "session" / "hyperloom_source_commit.txt",
-        start.parent.parent / "hyperloom_source_commit.txt",
-    ]
-    for sha_path in candidates:
-        if not sha_path.exists():
-            continue
-        try:
-            sha = sha_path.read_text(encoding="utf-8").strip()
-        except Exception:
-            continue
-        # Accept only SHA-shaped strings; never trust a corrupted file.
-        if 7 <= len(sha) <= 80 and all(c in "0123456789abcdef" for c in sha.lower()):
-            return sha
-
-    # Fallback: CI runner env. HYPERLOOM_SOURCE_REF is the pinned commit
-    # (preferred); GITHUB_SHA is the unconditional fallback.
-    for env_var in ("HYPERLOOM_SOURCE_REF", "GITHUB_SHA"):
-        env_sha = (os.environ.get(env_var) or "").strip()
-        if 7 <= len(env_sha) <= 80 and all(c in "0123456789abcdef" for c in env_sha.lower()):
-            return env_sha
-    return ""
-
-
-def _backfill_ci_metrics_file(path: Path, rec: SubmissionRecord) -> None:
-    """Backfill task metadata (model, image, hyperloom_commit, category,
-    sandbox_duration_seconds) into ci_metrics.json / session_breakdown.json /
-    manifest.json so each artifact is self-describing.
-
-    Writes the right shape per filename: ci_metrics.json → flat top-level;
-    session_breakdown.json → under session_meta; manifest.json → flat top-level
-    (V2 cli schema; category/duration are extra keys V2 ignores on re-read).
-
-    Args:
-        path (Path): Target JSON file (ci_metrics / session_breakdown /
-            manifest); other names and missing files are no-ops.
-        rec (SubmissionRecord): Record supplying the metadata to backfill.
-    """
-    if path.name not in ("ci_metrics.json", "session_breakdown.json", "manifest.json"):
-        return
-    if not path.exists():
-        return
-    try:
-        data = json.loads(path.read_text(encoding="utf-8"))
-    except Exception:
-        return
-    if not isinstance(data, dict):
-        return
-
-    changed = False
-    detected = rec.detected or {}
-    image = detected.get("image") or ""
-    hyperloom_sha = _find_hyperloom_commit_sha(path)
-    image_tag = image.split("/")[-1] if image else ""
-
-    if path.name == "ci_metrics.json":
-        for key, value in {
-            "model": rec.model,
-            "task_id": rec.task_id,
-            "claw_session_id": rec.claw_session_id,
-        }.items():
-            if value and not data.get(key):
-                data[key] = value
-                changed = True
-        for key in ("framework", "tp"):
-            if detected.get(key) is not None and data.get(key) is None:
-                data[key] = detected.get(key)
-                changed = True
-        if image and not data.get("image"):
-            data["image"] = image
-            changed = True
-        if hyperloom_sha and not data.get("hyperloom_commit"):
-            data["hyperloom_commit"] = hyperloom_sha
-            changed = True
-        if rec.category and not data.get("category"):
-            data["category"] = rec.category
-            changed = True
-        if rec.sandbox_duration_seconds is not None and not data.get("sandbox_duration_seconds"):
-            data["sandbox_duration_seconds"] = rec.sandbox_duration_seconds
-            changed = True
-
-    elif path.name == "session_breakdown.json":
-        # Keyed by a `session_meta` sub-dict; only write empty fields so we don't
-        # overwrite what the V2 collectors filled in.
-        meta = data.get("session_meta")
-        if not isinstance(meta, dict):
-            meta = {}
-            data["session_meta"] = meta
-        if image and not meta.get("image"):
-            meta["image"] = image
-            changed = True
-        if image_tag and not meta.get("image_id"):
-            meta["image_id"] = image_tag
-            changed = True
-        if hyperloom_sha and not meta.get("code_revision"):
-            meta["code_revision"] = hyperloom_sha
-            changed = True
-        if rec.sandbox_duration_seconds is not None and not meta.get("session_duration_seconds"):
-            meta["session_duration_seconds"] = rec.sandbox_duration_seconds
-            changed = True
-        # `category` isn't in the schema but unknown fields are tolerated.
-        if rec.category and not meta.get("category"):
-            meta["category"] = rec.category
-            changed = True
-
-    elif path.name == "manifest.json":
-        # V2 cli schema: flat top-level keys, often null when the sandbox didn't
-        # set HYPERLOOM_IMAGE or git rev-parse failed; backfill from the CI side.
-        for key, value in {
-            "model_name": rec.model,
-            "claw_session_id": rec.claw_session_id,
-        }.items():
-            if value and not data.get(key):
-                data[key] = value
-                changed = True
-        for key in ("framework", "tp"):
-            if detected.get(key) is not None and not data.get(key):
-                data[key] = detected.get(key)
-                changed = True
-        if image and not data.get("image"):
-            data["image"] = image
-            changed = True
-        if hyperloom_sha and not data.get("code_revision"):
-            data["code_revision"] = hyperloom_sha
-            changed = True
-        if rec.category and not data.get("category"):
-            data["category"] = rec.category
-            changed = True
-        if rec.sandbox_duration_seconds is not None and not data.get("sandbox_duration_seconds"):
-            data["sandbox_duration_seconds"] = rec.sandbox_duration_seconds
-            changed = True
-
-    if changed:
-        path.write_text(json.dumps(data, indent=2), encoding="utf-8")
-
-
-def _backfill_wekafs_in_place(rec: SubmissionRecord) -> int:
-    """Reverse-write audit fields back into the wekafs SOURCE files so operators
-    see them under /wekafs/users/<uid>/<sess>/ without GHA artifact zips
-    (image/category/duration are SaFE-side facts the agent/V2 cli never had).
-
-    Match (mirrors Stage B): exact `model` field, else conservative session-dir
-    match; only sessions modified in the last 24h. Updates ci_metrics.json,
-    manifest.json, session_breakdown[_v2].json across subdirs via
-    _backfill_ci_metrics_file. No-op when wekafs isn't mounted.
-
-    Args:
-        rec (SubmissionRecord): Record used to match sessions and supply the
-            metadata to backfill.
-
-    Returns:
-        int: Number of files updated; ``0`` when wekafs is unmounted or nothing
-        matched.
-    """
-    nfs_root = os.environ.get("NFS_ROOT", "/wekafs")
-    users_root = os.path.join(nfs_root, "users")
-    if not os.path.isdir(users_root):
-        return 0
-    target = _norm_token((rec.model or "").split("/")[-1])
-    if not target:
-        return 0
-
-    fresh_cutoff = time.time() - 24 * 3600
-    targets = ("ci_metrics.json", "manifest.json", "session_breakdown.json", "session_breakdown_v2.json")
-    subdirs = ("", "phase10_report", "results", "v2_session")
-
-    n = 0
-
-    def _session_has_matching_json(sess_path: str) -> bool:
-        """Return whether a session dir holds JSON matching the target model.
-
-        Args:
-            sess_path: Session directory to scan.
-
-        Returns:
-            ``True`` if any target JSON file under the known subdirs has a
-            model field matching ``rec``; otherwise ``False``.
-        """
-        for sub in subdirs:
-            base = os.path.join(sess_path, sub) if sub else sess_path
-            if not os.path.isdir(base):
-                continue
-            for fn in targets:
-                p = Path(base) / fn
-                if not p.is_file():
-                    continue
-                try:
-                    d = json.loads(p.read_text(encoding="utf-8"))
-                except Exception:
-                    continue
-                mf = _json_model_field(d) if isinstance(d, dict) else ""
-                if mf and _record_model_field_matches(rec, mf):
-                    return True
-        return False
-
-    def _backfill_files(sess_path: str) -> int:
-        """Backfill the model field into target JSON files in a session dir.
-
-        Args:
-            sess_path: Session directory whose target JSON files should be
-                updated.
-
-        Returns:
-            The number of files updated.
-        """
-        updated = 0
-        for sub in subdirs:
-            base = os.path.join(sess_path, sub) if sub else sess_path
-            if not os.path.isdir(base):
-                continue
-            for fn in targets:
-                p = Path(base) / fn
-                if not p.is_file():
-                    continue
-                try:
-                    before = p.read_bytes()
-                    _backfill_ci_metrics_file(p, rec)
-                    if p.read_bytes() != before:
-                        updated += 1
-                        log.info("[task %s] wekafs backfill: %s", rec.task_id, p)
-                except Exception as e:
-                    log.warning("[task %s] wekafs backfill failed for %s: %s", rec.task_id, p, e)
-        return updated
-
-    for uid_dir in os.listdir(users_root):
-        uid_path = os.path.join(users_root, uid_dir)
-        if not os.path.isdir(uid_path):
-            continue
-        for sess in os.listdir(uid_path):
-            sess_path = os.path.join(uid_path, sess)
-            if not os.path.isdir(sess_path):
-                continue
-            try:
-                if os.path.getmtime(sess_path) < fresh_cutoff:
-                    continue
-            except OSError:
-                continue
-            # Confirm session ownership: `model` field match (strongest), else
-            # the conservative session-dir-name heuristic.
-            matched = False
-            for sub in subdirs:
-                ci = (
-                    os.path.join(sess_path, sub, "ci_metrics.json")
-                    if sub
-                    else os.path.join(sess_path, "ci_metrics.json")
-                )
-                if not os.path.isfile(ci):
-                    continue
-                try:
-                    d = json.loads(Path(ci).read_text(encoding="utf-8"))
-                except Exception:
-                    continue
-                mf = str(d.get("model") or d.get("model_name") or "")
-                if mf and _record_model_field_matches(rec, mf):
-                    matched = True
-                    break
-            if not matched and _session_has_matching_json(sess_path):
-                matched = True
-            if not matched and _record_matches_session_dir(rec, sess):
-                matched = True
-            if not matched:
-                continue
-            n += _backfill_files(sess_path)
-
-        # Current layout: /wekafs/users/<uid>/<model-basename>/<YYYYmmddTHHMMSSZ>/.
-        # A deleted ci_metrics.json must not block manifest.json backfill.
-        if rec.safe_user_id and uid_dir != rec.safe_user_id:
-            continue
-        for model_dir_name in _candidate_model_dir_names(rec):
-            model_dir = os.path.join(uid_path, model_dir_name)
-            if not os.path.isdir(model_dir):
-                continue
-            try:
-                ts_entries = sorted(os.listdir(model_dir), reverse=True)
-            except Exception:
-                continue
-            for ts_entry in ts_entries:
-                sess_path = os.path.join(model_dir, ts_entry)
-                if not os.path.isdir(sess_path):
-                    continue
-                try:
-                    if os.path.getmtime(sess_path) < fresh_cutoff:
-                        continue
-                except OSError:
-                    continue
-                ts = _session_timestamp_from_path(ts_entry)
-                matched = _record_has_task_window(rec) and _timestamp_in_task_window(ts, rec)
-                if not matched and _session_has_matching_json(sess_path):
-                    matched = True
-                if not matched:
-                    continue
-                n += _backfill_files(sess_path)
-    return n
-
-
-def _record_matches_session_dir(rec: SubmissionRecord, sess_name: str) -> bool:
-    """Conservative directory-name match for the /wekafs/users fallback. Prefer
-    the displayName slug; fall back to basename with strict-term guards to avoid
-    cross-wiring adjacent repos (Qwen2.5 vs -AWQ, Nano vs Super, bnb, etc.).
-
-    Args:
-        rec (SubmissionRecord): Record providing the model id / display name.
-        sess_name (str): Candidate session directory name.
-
-    Returns:
-        bool: True when ``sess_name`` conservatively matches the record.
-    """
-    sess_norm = _norm_token(sess_name)
-    display = rec.display_name or ""
-    if display and _norm_token(display) in sess_norm:
-        return True
-
-    model = rec.model or ""
-    base = model.split("/")[-1]
-    base_norm = _norm_token(base)
-    if not base_norm or base_norm not in sess_norm:
-        return False
-
-    identity = f"{model} {display}".lower()
-    lower_dir = sess_name.lower()
-    strict_terms = ("awq", "gptq", "bnb", "4bit", "abliterated", "geneticlemonade", "nano", "super")
-    for term in strict_terms:
-        in_identity = term in identity
-        in_dir = term in lower_dir
-        if in_identity != in_dir:
-            return False
-    return True
-
-
-def _record_model_field_matches(rec: SubmissionRecord, model_field: str) -> bool:
-    """Compare a JSON model field with a SubmissionRecord conservatively.
-
-    Args:
-        rec (SubmissionRecord): Record supplying the allowed model names.
-        model_field (str): The model id read from a result JSON.
-
-    Returns:
-        bool: True when ``model_field`` normalizes to one of the record's
-        allowed names.
-    """
-    observed = _norm_token(str(model_field or ""))
-    if not observed:
-        return False
-    allowed = {
-        _norm_token((rec.model or "").split("/")[-1]),
-        _norm_token((rec.model or "").replace("/", "-")),
-        _norm_token((rec.model_path or "").rstrip("/\\").split("/")[-1]),
-        _norm_token(rec.display_name or ""),
-    }
-    allowed.discard("")
-    return observed in allowed or _norm_token(str(model_field).split("/")[-1]) in allowed
-
-
-def _candidate_model_dir_names(rec: SubmissionRecord) -> list[str]:
-    """Derive plausible per-model directory basenames for a record.
-
-    Args:
-        rec (SubmissionRecord): The record supplying model path / id / display
-            name candidates.
-
-    Returns:
-        list[str]: De-duplicated basename candidates, in priority order.
-    """
-    names: list[str] = []
-    for value in (
-        rec.model_path or "",
-        (rec.model or "").replace("/", "-"),
-        (rec.model or "").split("/")[-1],
-        rec.display_name or "",
-    ):
-        name = str(value or "").strip().rstrip("/\\").split("/")[-1]
-        if name and name not in names:
-            names.append(name)
-    return names
-
-
-def _json_positive_perf(data: dict) -> bool:
-    """Report whether a breakdown JSON has positive baseline AND optimized perf.
-
-    Args:
-        data (dict): A parsed session-breakdown/metrics JSON object.
-
-    Returns:
-        bool: True only if at least one positive baseline throughput and one
-            positive optimized throughput are present.
-    """
-
-    def positive(value: object) -> bool:
-        """Return True for a positive, non-bool numeric value.
-
-        Args:
-            value (object): Candidate value to test.
-
-        Returns:
-            bool: True if ``value`` is an ``int``/``float`` (not ``bool``)
-                greater than zero.
-        """
-        return isinstance(value, (int, float)) and not isinstance(value, bool) and value > 0
-
-    baseline = data.get("baseline") if isinstance(data.get("baseline"), dict) else {}
-    final = data.get("final") if isinstance(data.get("final"), dict) else {}
-    best = data.get("best") if isinstance(data.get("best"), dict) else {}
-    base_values = (
-        data.get("baseline_tput"),
-        data.get("baseline_throughput"),
-        data.get("tok_per_gpu_baseline"),
-        baseline.get("throughput_tok_s_per_gpu"),
-        baseline.get("output_throughput"),
-    )
-    opt_values = (
-        data.get("best_tput"),
-        data.get("optimized_throughput"),
-        data.get("tok_per_gpu_optimized"),
-        final.get("throughput_tok_s_per_gpu"),
-        final.get("output_throughput"),
-        best.get("throughput_tok_s_per_gpu"),
-        best.get("output_throughput"),
-    )
-    return any(positive(v) for v in base_values) and any(positive(v) for v in opt_values)
-
-
-def _json_model_field(data: dict) -> str:
-    """Extract the model id from a breakdown/metrics JSON, trying known keys.
-
-    Args:
-        data (dict): A parsed session-breakdown/metrics JSON object.
-
-    Returns:
-        str: The first non-empty model field found, or ``""``.
-    """
-    workload = data.get("workload") if isinstance(data.get("workload"), dict) else {}
-    session = data.get("session") if isinstance(data.get("session"), dict) else {}
-    meta = data.get("session_meta") if isinstance(data.get("session_meta"), dict) else {}
-    for value in (
-        data.get("model"),
-        data.get("model_name"),
-        workload.get("model_name"),
-        workload.get("model"),
-        session.get("model"),
-        meta.get("model"),
-    ):
-        if value not in (None, ""):
-            return str(value)
-    return ""
-
-
-def _json_claw_session_id(data: dict) -> str:
-    """Extract the Claw session id from a breakdown/metrics JSON.
-
-    Args:
-        data (dict): A parsed session-breakdown/metrics JSON object.
-
-    Returns:
-        str: The first non-empty ``claw_session_id`` found, or ``""``.
-    """
-    session = data.get("session") if isinstance(data.get("session"), dict) else {}
-    meta = data.get("session_meta") if isinstance(data.get("session_meta"), dict) else {}
-    for value in (
-        data.get("claw_session_id"),
-        session.get("claw_session_id"),
-        meta.get("claw_session_id"),
-    ):
-        if value not in (None, ""):
-            return str(value)
-    return ""
-
-
-def _resolve_record_claw_session_id(
-    safe: SafeOptimizeClient,
-    rec: SubmissionRecord,
-    last_task: dict | None = None,
-) -> str | None:
-    """Resolve a Claw session id for a submitted record.
-
-    SaFE's terminal task payload can occasionally omit ``clawSessionId`` even
-    though a subsequent task GET has it. Prefer the freshest terminal payload,
-    but fall back to the cached resolver before leaving the manifest without a
-    Claw id; otherwise Pulse classifies the dispatch as ``no_claw_session``.
-    """
-    for value in (
-        (last_task or {}).get("clawSessionId"),
-        rec.claw_session_id,
-    ):
-        sid = str(value or "").strip()
-        if sid:
-            return sid
-    if rec.task_id:
-        try:
-            sid = safe._claw_session_id_for(rec.task_id)
-            if sid:
-                return sid
-        except Exception as e:
-            log.debug("[task %s] clawSessionId fallback lookup failed: %s", rec.task_id, e)
-    return None
-
-
-def _env_truthy(name: str) -> bool:
-    """Report whether an environment variable is set to a truthy value.
-
-    Args:
-        name (str): The environment variable name.
-
-    Returns:
-        bool: True if the value is one of 1/true/yes/y/on (case-insensitive).
-    """
-    return (os.environ.get(name) or "").strip().lower() in {
-        "1",
-        "true",
-        "yes",
-        "y",
-        "on",
-    }
-
-
-def _copy_session_tree(src_dir: str, dst_dir: Path) -> int:
-    """Copy an entire persisted session directory into ``dst_dir`` (existing files
-    untouched). Returns the number of files copied.
-
-    Args:
-        src_dir (str): Source session directory to copy from.
-        dst_dir (Path): Destination directory (created if absent).
-
-    Returns:
-        int: Number of files copied.
-    """
-    copied = 0
-    dst_dir.mkdir(parents=True, exist_ok=True)
-    for root, dirnames, filenames in os.walk(src_dir):
-        dirnames[:] = [d for d in dirnames if d not in {".git", "__pycache__", ".pytest_cache"}]
-        rel_root = os.path.relpath(root, src_dir)
-        out_root = dst_dir if rel_root == "." else dst_dir / rel_root
-        out_root.mkdir(parents=True, exist_ok=True)
-        for fname in filenames:
-            src = os.path.join(root, fname)
-            dst = out_root / fname
-            if dst.exists():
-                continue
-            try:
-                shutil.copy2(src, dst)
-            except Exception as e:
-                log.warning("full-session copy failed %s -> %s: %s", src, dst, e)
-                continue
-            copied += 1
-    return copied
-
-
-def _nfs_fallback_collect(
-    rec: SubmissionRecord,
-    artifacts_dir: Path,
-    copy_full_session: bool = False,
-    current_session_hints: set[str] | None = None,
-) -> int:
-    """Scan NFS result directories for files matching this model, used when the
-    SaFE artifact API returns nothing (V2 writes under /wekafs/users/<uid>/...).
-
-    Two stages: A. legacy canonical CI dirs, matched by dir name; B. per-user
-    session dirs, matched by each ci_metrics.json's `model` field (more accurate
-    than dir-name fuzzy match). Mirrors ci/orchestrator.py's fallback.
-    Returns count of files copied.
-
-    Args:
-        rec (SubmissionRecord): Record used to match sessions and record
-            collected artifacts.
-        artifacts_dir (Path): Local artifacts root the files are copied into.
-        copy_full_session (bool): Also copy the full matched session tree.
-        current_session_hints (set[str] | None): Session-timestamp hints used to
-            scope the scan.
-
-    Returns:
-        int: Number of files copied.
-    """
-    current_session_hints = set(current_session_hints or set())
-    nfs_root = os.environ.get("NFS_ROOT", "/wekafs")
-    model_basename = (rec.model or "").split("/")[-1]
-    model_short = model_basename.lower().replace("-", "").replace("_", "").replace(".", "")
-    if not model_short or not rec.task_id:
-        return 0
-
-    task_dir = artifacts_dir / rec.task_id
-    task_dir.mkdir(parents=True, exist_ok=True)
-
-    copied = 0
-
-    has_task_scoped_model_dir = bool(rec.safe_user_id and _record_has_task_window(rec))
-    if not current_session_hints and not has_task_scoped_model_dir:
-        log.info(
-            "[task %s] NFS fallback skipped: no current-session "
-            "timestamp hints from SaFE artifacts or task time window",
-            rec.task_id,
-        )
-        return copied
-
-    # ── Stage A: legacy canonical dirs (matched by current timestamp) ───────
-    legacy_scan_dirs = [
-        f"{nfs_root}/hyperloom-results",
-        f"{nfs_root}/results/ci",
-        f"{nfs_root}/inference-optimization/results",
-    ]
-    for scan_dir in legacy_scan_dirs:
-        if not os.path.isdir(scan_dir):
-            continue
-        try:
-            entries = sorted(os.listdir(scan_dir), reverse=True)
-        except Exception:
-            continue
-        for entry in entries:
-            if not _path_has_session_hint(entry, current_session_hints):
-                continue
-            entry_clean = entry.lower().replace("-", "").replace("_", "").replace(".", "")
-            if model_short not in entry_clean:
-                continue
-            candidate = os.path.join(scan_dir, entry)
-            if not os.path.isdir(candidate):
-                continue
-            for suffix in _KEY_RESULT_SUFFIXES:
-                for root, _, fnames in os.walk(candidate):
-                    for fn in fnames:
-                        if not fn.endswith(suffix):
-                            continue
-                        src = os.path.join(root, fn)
-                        dst = task_dir / fn
-                        if dst.exists():
-                            continue
-                        try:
-                            shutil.copy2(src, dst)
-                            _backfill_ci_metrics_file(dst, rec)
-                            _record_artifact_source(
-                                rec,
-                                dst,
-                                "nfs_legacy",
-                                source_path=src,
-                                session_dir=candidate,
-                            )
-                        except Exception as e:
-                            log.warning("[task %s] NFS legacy copy %s -> %s failed: %s", rec.task_id, src, dst, e)
-                            continue
-                        rec.artifact_files.append(str(dst))
-                        rec.artifact_count += 1
-                        copied += 1
-                        log.info("[task %s] NFS legacy: copied %s -> %s", rec.task_id, src, dst)
-            if copied:
-                break
-        if copied:
-            break
-    if copied:
-        return copied
-
-    # ── Stage B: /wekafs/users/<uid>/<session>/...  matched by timestamp ────
-    users_root = f"{nfs_root}/users"
-    if not os.path.isdir(users_root):
-        return copied
-
-    # Primary match: JSON model fields when present. Secondary: conservative
-    # session-dir match (needed for runs that persist JSON without a `model` field).
-    target = _norm_token(model_basename)
-    if not target:
-        return copied
-
-    def _model_field_matches(model_field: str) -> bool:
-        """Check a JSON ``model`` field against the record's expected names.
-
-        Args:
-            model_field (str): The model id read from a candidate result file.
-
-        Returns:
-            bool: True if it normalizes to one of the record's allowed names.
-        """
-        observed = _norm_token(model_field)
-        allowed = {
-            target,
-            _norm_token((rec.model or "").replace("/", "-")),
-            _norm_token((rec.model_path or "").rstrip("/\\").split("/")[-1]),
-        }
-        allowed.discard("")
-        return observed in allowed or _norm_token(model_field.split("/")[-1]) in allowed
-
-    def _consider_result_file(path: str, session_dir: str, score_base: int) -> None:
-        """Score a candidate result file and append it to ``candidates``.
-
-        Validates the file's model/session/claw fields against the record and,
-        when it matches, records a ``(score, mtime, path, session_dir)`` tuple
-        for later best-match selection.
-
-        Args:
-            path (str): Path to a candidate result JSON file.
-            session_dir (str): The session directory containing the file.
-            score_base (int): Base score reflecting match-source confidence.
-        """
-        if not os.path.isfile(path):
-            return
-        try:
-            with open(path, encoding="utf-8") as f:
-                data = json.load(f)
-        except Exception:
-            return
-        if not isinstance(data, dict):
-            return
-
-        model_field = _json_model_field(data)
-        if model_field:
-            if not _model_field_matches(model_field):
-                return
-            score = score_base + 100
-        else:
-            session_name = os.path.basename(session_dir.rstrip("/"))
-            parent_name = os.path.basename(os.path.dirname(session_dir.rstrip("/")))
-            if not (_record_matches_session_dir(rec, session_name) or _record_matches_session_dir(rec, parent_name)):
-                return
-            score = score_base + 60
-
-        claw = _json_claw_session_id(data)
-        if rec.claw_session_id and claw:
-            if rec.claw_session_id != claw:
-                return
-            score += 30
-        elif claw and not rec.claw_session_id:
-            rec.claw_session_id = claw
-            score += 20
-        if _json_positive_perf(data):
-            score += 10
-        else:
-            log.info(
-                "[task %s] candidate %s matched but has no positive throughput (will use only if no better match)",
-                rec.task_id,
-                path,
-            )
-        try:
-            mtime = os.path.getmtime(path)
-        except OSError:
-            return
-        candidates.append((score, mtime, path, session_dir))
-
-    candidates: list[tuple[int, float, str, str]] = []  # (score, mtime, marker_path, session_dir)
-    uid_dirs = [rec.safe_user_id] if rec.safe_user_id else os.listdir(users_root)
-    for uid_dir in uid_dirs:
-        if not uid_dir:
-            continue
-        uid_path = os.path.join(users_root, uid_dir)
-        if not os.path.isdir(uid_path):
-            continue
-        for sess in os.listdir(uid_path):
-            sess_path = os.path.join(uid_path, sess)
-            if not os.path.isdir(sess_path):
-                continue
-            if not _path_has_session_hint(sess_path, current_session_hints):
-                continue
-            for sub in ("", "phase10_report", "results"):
-                ci_path = (
-                    os.path.join(sess_path, sub, "ci_metrics.json")
-                    if sub
-                    else os.path.join(sess_path, "ci_metrics.json")
-                )
-                _consider_result_file(ci_path, sess_path, 0)
-
-        # Current layout: /wekafs/users/<uid>/<model-basename>/<YYYYmmddTHHMMSSZ>/.
-        # With no artifact-derived hint, accept timestamp dirs only under this
-        # exact user id + model dir and inside the task's startedAt/finishedAt window.
-        if rec.safe_user_id and uid_dir == rec.safe_user_id:
-            for model_dir_name in _candidate_model_dir_names(rec):
-                model_dir = os.path.join(uid_path, model_dir_name)
-                if not os.path.isdir(model_dir):
-                    continue
-                try:
-                    ts_entries = sorted(os.listdir(model_dir), reverse=True)
-                except Exception:
-                    continue
-                for ts_entry in ts_entries:
-                    session_dir = os.path.join(model_dir, ts_entry)
-                    if not os.path.isdir(session_dir):
-                        continue
-                    ts = _session_timestamp_from_path(ts_entry)
-                    if current_session_hints:
-                        if not _path_has_session_hint(session_dir, current_session_hints):
-                            continue
-                        score_base = 40
-                    else:
-                        if not _timestamp_in_task_window(ts, rec):
-                            continue
-                        score_base = 30
-                    for sub, filename in (
-                        ("", "ci_metrics.json"),
-                        ("phase10_report", "ci_metrics.json"),
-                        ("results", "ci_metrics.json"),
-                        ("", "session_breakdown.json"),
-                        ("phase10_report", "session_breakdown.json"),
-                        ("v2_session", "session_breakdown.json"),
-                    ):
-                        result_path = (
-                            os.path.join(session_dir, sub, filename) if sub else os.path.join(session_dir, filename)
-                        )
-                        _consider_result_file(result_path, session_dir, score_base)
-
-    if not candidates:
-        log.info(
-            "[task %s] no /wekafs/users candidate matched model=%s (user_id=%s, hints=%s, task_window=%s)",
-            rec.task_id,
-            model_basename,
-            rec.safe_user_id or "?",
-            ",".join(sorted(current_session_hints)) or "-",
-            "yes" if _record_has_task_window(rec) else "no",
-        )
-        return copied
-
-    # Highest-confidence, then freshest (same model may be re-run).
-    candidates.sort(reverse=True)
-    _score, _mtime, best_ci, best_sess = candidates[0]
-    log.info(
-        "[task %s] NFS user-session match: %s (from %d candidate(s), session=%s)",
-        rec.task_id,
-        best_ci,
-        len(candidates),
-        best_sess,
-    )
-
-    # Copy ci_metrics + any optimization_report.md flat under task_dir/ (the
-    # shape build_summary.py expects).
-    targets = [best_ci]
-    for cand in [
-        os.path.join(best_sess, "optimization_report.md"),
-        os.path.join(best_sess, "phase10_report", "optimization_report.md"),
-        os.path.join(best_sess, "reports", "final.md"),
-    ]:
-        if os.path.isfile(cand):
-            targets.append(cand)
-            break
-    # Optional audit artifact, when the agent emitted it.
-    for cand in [
-        os.path.join(best_sess, "session_breakdown.json"),
-        os.path.join(best_sess, "phase10_report", "session_breakdown.json"),
-    ]:
-        if os.path.isfile(cand):
-            targets.append(cand)
-            break
-
-    for src in targets:
-        dst_name = "optimization_report.md" if src.endswith("final.md") else os.path.basename(src)
-        dst = task_dir / dst_name
-        if dst.exists():
-            continue
-        try:
-            shutil.copy2(src, dst)
-            _backfill_ci_metrics_file(dst, rec)
-            _record_artifact_source(
-                rec,
-                dst,
-                "nfs_user_session",
-                source_path=src,
-                session_dir=best_sess,
-            )
-        except Exception as e:
-            log.warning("[task %s] NFS user-session copy %s -> %s failed: %s", rec.task_id, src, dst, e)
-            continue
-        rec.artifact_files.append(str(dst))
-        rec.artifact_count += 1
-        copied += 1
-        log.info("[task %s] NFS user-session: copied %s -> %s", rec.task_id, src, dst)
-
-    if copy_full_session:
-        session_dst = task_dir / "session"
-        n_full = _copy_session_tree(best_sess, session_dst)
-        if n_full:
-            copied += n_full
-            rec.artifact_count += n_full
-            rec.artifact_files.append(str(session_dst))
-            log.info(
-                "[task %s] NFS full-session: copied %d file(s) %s -> %s", rec.task_id, n_full, best_sess, session_dst
-            )
-        else:
-            log.info("[task %s] NFS full-session: no new files copied from %s", rec.task_id, best_sess)
-
-    return copied
-
-
-def wait_and_collect_one(
-    safe: SafeOptimizeClient,
-    rec: SubmissionRecord,
-    artifacts_dir: Path,
-    task_timeout_min: int,
-    poll_s: int,
-    collect: bool,
-    all_artifacts: bool,
-) -> SubmissionRecord:
-    """Wait for one task to finish, then optionally download its artifacts:
-    (1) SaFE artifacts API, then (2) NFS fallback when session_breakdown.json
-    (the CI delivery contract) is still missing.
-
-    Args:
-        safe (SafeOptimizeClient): Client used to poll and download artifacts.
-        rec (SubmissionRecord): Record to update with status and artifacts.
-        artifacts_dir (Path): Local artifacts root for this task.
-        task_timeout_min (int): Max minutes to wait for the task.
-        poll_s (int): Polling interval in seconds.
-        collect (bool): When False, skip artifact collection.
-        all_artifacts (bool): When True, download every artifact / full session.
-
-    Returns:
-        SubmissionRecord: The same ``rec``, mutated in place with results.
-    """
-    if not rec.task_id:
-        return rec  # nothing to wait for (skipped/failed during submit)
-
-    final_status, last_task = safe.wait_task_done(rec.task_id, timeout_min=task_timeout_min, poll_s=poll_s)
-    rec.final_status = final_status
-    rec.final_phase = last_task.get("currentPhase")
-    rec.final_message = (last_task.get("message") or "")[:500] or None
-    rec.claw_session_id = _resolve_record_claw_session_id(safe, rec, last_task)
-    rec.model_path = (last_task.get("modelPath") or "").strip() or rec.model_path
-    rec.safe_user_id = (last_task.get("userId") or "").strip() or rec.safe_user_id
-    rec.safe_started_at = (last_task.get("startedAt") or "").strip() or rec.safe_started_at
-    rec.safe_finished_at = (last_task.get("finishedAt") or "").strip() or rec.safe_finished_at
-    rec.sandbox_duration_seconds = _sandbox_duration_seconds(last_task)
-    if rec.claw_session_id:
-        log.info(
-            "[task %s] clawSessionId=%s duration=%ss",
-            rec.task_id,
-            rec.claw_session_id,
-            rec.sandbox_duration_seconds if rec.sandbox_duration_seconds is not None else "?",
-        )
-
-    if not collect:
-        rec.ci_status = "Not collected"
-        rec.delivery_reason = "artifact collection disabled"
-        return rec
-
-    # Stage 1: SaFE artifacts API (most reliable when the agent copied files to
-    # /workspace/hyperloom). Retry — terminal status can beat Claw's file index.
-    items = []
-    wanted = []
-    for attempt in range(3):
-        try:
-            items = safe.list_artifacts(rec.task_id)
-        except Exception as e:
-            log.warning("[task %s] list_artifacts attempt %d failed: %s", rec.task_id, attempt + 1, e)
-            items = []
-        wanted = [it for it in items if _is_wanted_artifact(it.get("path", ""), all_artifacts)]
-        wanted_paths = [it.get("path", "").lower() for it in wanted]
-        # session_breakdown.json is the CI delivery contract (cli.finally always
-        # writes it, even on abort), so retry only until it shows up.
-        has_safe_breakdown = any(p.endswith("session_breakdown.json") for p in wanted_paths)
-        if has_safe_breakdown:
-            break
-        if attempt < 2:
-            log.info(
-                "[task %s] safe artifacts missing session_breakdown.json on attempt %d; retrying",
-                rec.task_id,
-                attempt + 1,
-            )
-            time.sleep(15)
-    log.info("[task %s] safe artifacts: %d total, %d to download", rec.task_id, len(items), len(wanted))
-    current_session_hints = _session_hints_from_artifact_items(items)
-    if current_session_hints:
-        log.info(
-            "[task %s] current session timestamp hints from artifacts: %s",
-            rec.task_id,
-            ", ".join(sorted(current_session_hints)),
-        )
-
-    if not has_safe_breakdown:
-        waited_session = _wait_for_nfs_session_delivery(
-            rec,
-            current_session_hints=current_session_hints,
-            poll_s=poll_s,
-        )
-        if waited_session:
-            # The SaFE artifact index may lag behind the agent's final writes.
-            # Re-list once after the grace wait before falling back to NFS.
-            try:
-                items = safe.list_artifacts(rec.task_id)
-                wanted = [it for it in items if _is_wanted_artifact(it.get("path", ""), all_artifacts)]
-                wanted_paths = [it.get("path", "").lower() for it in wanted]
-                current_session_hints.update(_session_hints_from_artifact_items(items))
-                log.info(
-                    "[task %s] safe artifacts after NFS grace wait: %d total, %d to download",
-                    rec.task_id,
-                    len(items),
-                    len(wanted),
-                )
-            except Exception as e:
-                log.warning("[task %s] post-grace list_artifacts failed: %s", rec.task_id, e)
-
-    task_dir = artifacts_dir / rec.task_id
-    rec.artifacts_dir = str(task_dir)
-    for it in wanted:
-        path = it.get("path", "")
-        if not path:
-            continue
-        local = _safe_local_path(artifacts_dir, rec.task_id, path)
-        try:
-            n = safe.download_artifact_to(rec.task_id, it, str(local))
-            _backfill_ci_metrics_file(local, rec)
-            _record_artifact_source(
-                rec,
-                local,
-                "safe_artifact_api",
-                remote_path=path,
-            )
-            rec.artifact_files.append(str(local))
-            rec.artifact_count += 1
-            log.info("[task %s] saved %s (%d bytes)", rec.task_id, path, n)
-        except Exception as e:
-            log.warning("[task %s] failed to download %s: %s", rec.task_id, path, e)
-
-    # Stage 2: NFS fallback when Stage 1 didn't deliver session_breakdown.json
-    # (the CI delivery contract). ci_metrics.json / optimization_report.md no
-    # longer gate the fallback.
-    has_breakdown = any(p.endswith("session_breakdown.json") for p in rec.artifact_files)
-    if not has_breakdown:
-        log.info("[task %s] missing session_breakdown.json — trying NFS fallback", rec.task_id)
-        copy_full_session = all_artifacts or _env_truthy("SAFE_OPTIMIZE_COPY_FULL_SESSION")
-        n_added = _nfs_fallback_collect(
-            rec,
-            artifacts_dir,
-            copy_full_session=copy_full_session,
-            current_session_hints=current_session_hints,
-        )
-        if n_added:
-            log.info("[task %s] NFS fallback added %d files", rec.task_id, n_added)
-        else:
-            log.info("[task %s] NFS fallback found nothing", rec.task_id)
-
-    _mark_record_delivery(rec)
-    if rec.ci_status:
-        log.info("[task %s] CI delivery status: %s (%s)", rec.task_id, rec.ci_status, rec.delivery_reason or "-")
-
-    # Stage 3: reverse-backfill audit fields into the wekafs SOURCE files so
-    # operators see them without the GHA artifact zip. No-op when wekafs unmounted.
-    if rec.artifact_count:
-        try:
-            n_wkfs = _backfill_wekafs_in_place(rec)
-            if n_wkfs:
-                log.info("[task %s] wekafs in-place backfill updated %d file(s)", rec.task_id, n_wkfs)
-        except Exception as e:
-            log.warning("[task %s] wekafs in-place backfill skipped due to %s: %s", rec.task_id, type(e).__name__, e)
-    else:
-        log.info("[task %s] wekafs in-place backfill skipped: no artifacts collected", rec.task_id)
-    try:
-        _write_artifact_sources(task_dir, rec)
-    except Exception as e:
-        log.warning("[task %s] artifact source metadata write skipped: %s", rec.task_id, e)
-    return rec
-
-
-def process_completion(
-    safe: SafeOptimizeClient,
-    records: list[SubmissionRecord],
-    artifacts_dir: Path,
-    task_timeout_min: int,
-    poll_s: int,
-    collect: bool,
-    all_artifacts: bool,
-    parallel: int,
-) -> None:
-    """Wait + collect for all submitted records, in parallel up to ``parallel``.
-
-    Args:
-        safe (SafeOptimizeClient): Client used to poll and download artifacts.
-        records (list[SubmissionRecord]): All submission records; only
-            ``submitted`` ones with a task id are awaited.
-        artifacts_dir (Path): Destination root for downloaded artifacts.
-        task_timeout_min (int): Max minutes to wait per task.
-        poll_s (int): Polling interval in seconds.
-        collect (bool): When True, collect artifacts after each task finishes.
-        all_artifacts (bool): When True, also copy full session trees.
-        parallel (int): Max concurrent wait/collect workers (<=1 runs serially).
-    """
-    pending = [r for r in records if r.status == "submitted" and r.task_id]
-    if not pending:
-        log.info("no submitted tasks to wait for")
-        return
-
-    artifacts_dir.mkdir(parents=True, exist_ok=True)
-    log.info(
-        "waiting for %d task(s) to finish (parallel=%d, timeout=%dm each)", len(pending), parallel, task_timeout_min
-    )
-
-    if parallel <= 1:
-        for rec in pending:
-            try:
-                wait_and_collect_one(safe, rec, artifacts_dir, task_timeout_min, poll_s, collect, all_artifacts)
-            except Exception as e:
-                log.exception("[task %s] unexpected wait/collect error", rec.task_id)
-                rec.final_status = rec.final_status or "Error"
-                rec.final_message = (rec.final_message or "") + f" | wait error: {e}"
-            finally:
-                if not rec.ci_status:
-                    _mark_record_delivery(rec)
-        return
-
-    from concurrent.futures import ThreadPoolExecutor, as_completed
-
-    with ThreadPoolExecutor(max_workers=parallel) as ex:
-        futures = {
-            ex.submit(
-                wait_and_collect_one, safe, rec, artifacts_dir, task_timeout_min, poll_s, collect, all_artifacts
-            ): rec
-            for rec in pending
-        }
-        for fut in as_completed(futures):
-            rec = futures[fut]
-            try:
-                fut.result()  # mutates rec in place
-            except Exception as e:
-                log.exception("[task %s] unexpected wait/collect error", rec.task_id)
-                rec.final_status = rec.final_status or "Error"
-                rec.final_message = (rec.final_message or "") + f" | wait error: {e}"
-            finally:
-                if not rec.ci_status:
-                    _mark_record_delivery(rec)
-
-
-# ── Manifest ────────────────────────────────────────────────────────────────────
-
-
-def write_manifest(
-    out_dir: Path,
-    records: list[SubmissionRecord],
-    base_url: str,
-    register_workspace: str,
-    submit_workspace: str,
-    volume: str,
-) -> None:
-    """Write the submission manifest as JSON and a markdown summary table.
-
-    Args:
-        out_dir (Path): Output directory (created if absent).
-        records (list[SubmissionRecord]): The submission records to serialize.
-        base_url (str): SaFE API base URL recorded in the manifest.
-        register_workspace (str): Workspace used for registration.
-        submit_workspace (str): Workspace used for submission.
-        volume (str): Storage volume name recorded in the manifest.
-    """
-    out_dir.mkdir(parents=True, exist_ok=True)
-    payload = {
-        "submitted_at": time.strftime("%Y-%m-%dT%H:%M:%SZ", time.gmtime()),
-        "api_url": base_url,
-        "register_workspace": register_workspace,
-        "submit_workspace": submit_workspace,
-        "volume": volume,
-        "records": [asdict(r) for r in records],
-    }
-    (out_dir / "submission_manifest.json").write_text(json.dumps(payload, indent=2))
-
-    md = [
-        "# SaFE Optimization Submission Manifest",
-        f"- API: `{base_url}`",
-        f"- Register workspace: `{register_workspace}`",
-        f"- Submit workspace: `{submit_workspace}`",
-        f"- Volume: `{volume}`",
-        f"- Submitted at: {payload['submitted_at']}",
-        "",
-        "| Pool | Model | Category | Image | Duration | Submit | Final | CI | Phase | Task ID | Display Name | Artifacts | Note |",
-        "|---|---|---|---|---|---|---|---|---|---|---|---|---|",
-    ]
-    for r in records:
-        # final_status is only meaningful when --wait-for-completion was on.
-        final = r.final_status or ("-" if r.status == "submitted" else "")
-        phase = "-" if r.final_phase is None else str(r.final_phase)
-        artifacts_cell = f"{r.artifact_count} files in `{r.artifacts_dir}`" if r.artifact_count else "-"
-        note_parts = []
-        if r.error:
-            note_parts.append(r.error)
-        if r.final_message:
-            note_parts.append(r.final_message)
-        note = " \\| ".join(note_parts).replace("|", "\\|")[:200]
-        pool_cell = "-"
-        if r.pool:
-            pool_id = r.pool.get("pool_id") or "-"
-            pool_idx = r.pool.get("pool_index")
-            batch_idx = r.pool.get("batch_index")
-            batch_size = r.pool.get("batch_size")
-            pool_cell = (
-                f"`{pool_id}`"
-                f"<br/>idx={pool_idx if pool_idx not in (None, '') else '-'}"
-                f"<br/>batch={batch_idx if batch_idx not in (None, '') else '-'}/"
-                f"{batch_size if batch_size not in (None, '') else '-'}"
-            )
-        # Image cell: tag suffix only for readability; full path is in JSON.
-        image_cell = "-"
-        image_full = (r.detected or {}).get("image", "") if r.detected else ""
-        if image_full:
-            image_cell = "`" + image_full.split("/")[-1] + "`"
-        # Duration cell: rounded minutes; ms-precision is in JSON.
-        duration_cell = "-"
-        if r.sandbox_duration_seconds is not None:
-            mins = r.sandbox_duration_seconds / 60.0
-            duration_cell = f"{mins:.1f}m"
-        category_cell = r.category or "-"
-        ci_cell = r.ci_status or ("Succeeded" if r.final_status == "Succeeded" else "-")
-        if r.ci_success and r.delivery_reason:
-            ci_cell = f"{ci_cell}<br/>{r.delivery_reason}"
-        md.append(
-            f"| {pool_cell} | `{r.model}` | {category_cell} | {image_cell} | {duration_cell} | "
-            f"{r.status} | {final or '-'} | {ci_cell} | {phase} | "
-            f"`{r.task_id or '-'}` | {r.display_name or '-'} | {artifacts_cell} | {note} |"
-        )
-    (out_dir / "submission_manifest.md").write_text("\n".join(md) + "\n")
-    log.info("manifest written to %s", out_dir)
+def _sync_artifact_facade_overrides() -> None:
+    """Keep legacy monkeypatches on optimize_submit.* visible to artifacts."""
+    _artifacts._wait_for_nfs_session_delivery = globals()["_wait_for_nfs_session_delivery"]
+
+
+def wait_and_collect_one(*args, **kwargs):
+    """Compatibility wrapper for tests/importers that monkeypatch facade hooks."""
+    _sync_artifact_facade_overrides()
+    return _artifacts.wait_and_collect_one(*args, **kwargs)
 
 
 # ── CLI ─────────────────────────────────────────────────────────────────────────
